redis::aio
Struct ConnectionManagerConfig

pub struct ConnectionManagerConfig { /* private fields */ }

The configuration for reconnect mechanism and request timing for the ConnectionManager

Implementations

impl ConnectionManagerConfig

pub fn new() -> Self

Creates a new instance of the options with nothing set

pub fn set_factor(self, factor: u64) -> ConnectionManagerConfig

A multiplicative factor that will be applied to the retry delay.

For example, using a factor of 1000 will make each delay in units of seconds.

pub fn set_max_delay(self, time: u64) -> ConnectionManagerConfig

Apply a maximum delay between connection attempts. The delay between attempts wonโt be longer than max_delay milliseconds.

pub fn set_exponent_base(self, base: u64) -> ConnectionManagerConfig

The resulting duration is calculated by taking the base to the n-th power, where n denotes the number of past attempts.

pub fn set_number_of_retries(self, amount: usize) -> ConnectionManagerConfig

number_of_retries times, with an exponentially increasing delay

pub fn set_response_timeout(self, duration: Duration) -> ConnectionManagerConfig

The new connection will time out operations after response_timeout has passed.

pub fn set_connection_timeout( self, duration: Duration, ) -> ConnectionManagerConfig

Each connection attempt to the server will time out after connection_timeout.

pub fn set_push_sender(self, sender: impl AsyncPushSender) -> Self

Sets sender sender for push values.

The sender can be a channel, or an arbitrary function that handles crate::PushInfo values. This will fail client creation if the connection isnโt configured for RESP3 communications via the crate::RedisConnectionInfo::protocol field. Setting this will mean that the connection manager actively listens to updates from the server, and so it will cause the manager to reconnect after a disconnection, even if the manager was unused at the time of the disconnect.

Examples

let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel();
let config = ConnectionManagerConfig::new().set_push_sender(tx);

let messages = Arc::new(Mutex::new(Vec::new()));
let config = ConnectionManagerConfig::new().set_push_sender(move |msg|{
    let Ok(mut messages) = messages.lock() else {
        return Err(redis::aio::SendError);
    };
    messages.push(msg);
    Ok(())
});

pub fn set_automatic_resubscription(self) -> Self

Configures the connection manager to automatically resubscribe to all pubsub channels after reconnecting.