Struct redis::aio::PubSub

pub struct PubSub<C = Pin<Box<dyn AsyncStream + Send + Sync>>>(_);

Represents a PubSub connection.

Implementations

impl<C> PubSub<C> where
C: Unpin + AsyncRead + AsyncWrite + Send,

pub async fn subscribe<T: ToRedisArgs>(&mut self, channel: T) -> RedisResult<()>

Subscribes to a new channel.

pub async fn psubscribe<T: ToRedisArgs>(
&mut self,
pchannel: T
) -> RedisResult<()>

Subscribes to a new channel with a pattern.

pub async fn unsubscribe<T: ToRedisArgs>(
&mut self,
channel: T
) -> RedisResult<()>

Unsubscribes from a channel.

pub async fn punsubscribe<T: ToRedisArgs>(
&mut self,
pchannel: T
) -> RedisResult<()>

Unsubscribes from a channel with a pattern.

pub fn on_message<'a>(&'a mut self) -> impl Stream<Item = Msg> + 'a

Returns Stream of Msgs from this PubSubs subscriptions.

The message itself is still generic and can be converted into an appropriate type through the helper methods on it.

pub fn into_on_message(self) -> impl Stream<Item = Msg>

Returns Stream of Msgs from this PubSubs subscriptions consuming it.

The message itself is still generic and can be converted into an appropriate type through the helper methods on it. This can be useful in cases where the stream needs to be returned or held by something other

pub async fn into_connection(self) -> Connection<C>

Exits from PubSub mode and converts PubSub into Connection.