redis::aio

Trait AsyncPushSender

pub trait AsyncPushSender:
    Send
    + Sync
    + 'static {
    // Required method
    fn send(&self, info: PushInfo) -> Result<(), SendError>;
}

A trait for sender parts of a channel that can be used for sending push messages from async connection.

Required Methods

fn send(&self, info: PushInfo) -> Result<(), SendError>

The sender must send without blocking, otherwise it will block the sending connection.