Trait redis::async::ConnectionLike

pub trait ConnectionLike: Sized {
    fn req_packed_command(self, cmd: Vec<u8>) -> RedisFuture<(Self, Value)>;

    fn req_packed_commands(
        self, 
        cmd: Vec<u8>, 
        offset: usize, 
        count: usize
    ) -> RedisFuture<(Self, Vec<Value>)>;

    fn get_db(&self) -> i64;
}

Required methods

fn req_packed_command(self, cmd: Vec<u8>) -> RedisFuture<(Self, Value)>

Sends an already encoded (packed) command into the TCP socket and reads the single response from it.

fn req_packed_commands(
self,
cmd: Vec<u8>,
offset: usize,
count: usize
) -> RedisFuture<(Self, Vec<Value>)>

Sends multiple already encoded (packed) command into the TCP socket and reads count responses from it. This is used to implement pipelining.

fn get_db(&self) -> i64

Returns the database this connection is bound to. Note that this information might be unreliable because it's initially cached and also might be incorrect if the connection like object is not actually connected.

Loading content...