redis

Module cluster_routing

Routing information for cluster commands.

Structs

Route

Defines the slot and the SlotAddr to which a command should be sent

Enums

AggregateOp

Numerical aggreagting operators.

LogicalAggregateOp

Logical bitwise aggregating operators.

MultipleNodeRoutingInfo

Defines which collection of nodes should receive a request

ResponsePolicy

Policy defining how to combine multiple responses into one.

RoutingInfo

Defines whether a request should be routed to a single node, or multiple ones.

SingleNodeRoutingInfo

Defines which single node should receive a request.

SlotAddr

What type of node should a request be routed to.

Traits

Routable

Objects that implement this trait define a request that can be routed by a cluster client to different nodes in the cluster.

Functions

command_for_multi_slot_indices

Takes a routable and an iterator of indices, which is assued to be created fromMultipleNodeRoutingInfo::MultiSlot, and returns a command with the arguments matching the indices.

get_slot

Returns the slot that matches key.