Enum redis::ErrorKind

pub enum ErrorKind {
    ResponseError,
    AuthenticationFailed,
    TypeError,
    ExecAbortError,
    BusyLoadingError,
    NoScriptError,
    InvalidClientConfig,
    ExtensionError(String),
    InternalIoError(IoError),
}

An enum of all error kinds.

Variants

ResponseError

The server generated an invalid response.

AuthenticationFailed

The authentication with the server failed.

TypeError

Operation failed because of a type mismatch.

ExecAbortError

A script execution was aborted.

BusyLoadingError

The server cannot response because it's loading a dump.

NoScriptError

A script that was requested does not actually exist.

InvalidClientConfig

An error that was caused because the parameter to the client were wrong.

ExtensionError

A error that is unknown to the library. This might be used by future redis servers or custom extensions.

InternalIoError

An IO error was encountered while talking to the server.