redis

Enum ExpireOption

pub enum ExpireOption {
    NONE,
    NX,
    XX,
    GT,
    LT,
}

Helper enum that is used to define option for the hash expire commands

Variants

NONE

NONE โ Set expiration regardless of the fieldโs current expiration.

NX

NX โ Only set expiration only when the field has no expiration.

XX

XX โ Only set expiration only when the field has an existing expiration.

GT

GT โ Only set expiration only when the new expiration is greater than current one.

LT

LT โ Only set expiration only when the new expiration is less than current one.