Function redis::cmd

pub fn cmd<'a>(name: &'a str) -> Cmd

Shortcut function to creating a command with a single argument.

The first argument of a redis command is always the name of the command which needs to be a string. This is the recommended way to start a command pipe.

redis::cmd("PING");