redis::streams

Enum XDelExStatusCode

#[non_exhaustive]
pub enum XDelExStatusCode {
    IdNotFound = -1,
    Deleted = 1,
    NotDeletedUnacknowledgedOrStillReferenced = 2,
}

Status codes returned by the XDELEX command

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

IdNotFound = -1

No entry with the given id exists in the stream

Deleted = 1

The entry was deleted from the stream

NotDeletedUnacknowledgedOrStillReferenced = 2

The entry was not deleted because it has either not been delivered to any consumer or still has references in the consumer groups’ Pending Entries List (PEL)