Struct redis::streams::StreamReadOptions

pub struct StreamReadOptions { /* private fields */ }

Builder options for xread_options command.

Implementations

impl StreamReadOptions

pub fn read_only(&self) -> bool

Indicates whether the command is participating in a group and generating ACKs

pub fn noack(self) -> Self

Sets the command so that it avoids adding the message to the PEL in cases where reliability is not a requirement and the occasional message loss is acceptable.

pub fn block(self, ms: usize) -> Self

Sets the block time in milliseconds.

pub fn count(self, n: usize) -> Self

Sets the maximum number of elements to return per stream.

pub fn group<GN: ToRedisArgs, CN: ToRedisArgs>( self, group_name: GN, consumer_name: CN ) -> Self

Sets the name of a consumer group associated to the stream.