Struct redis::Parser

pub struct Parser<T> {
    // some fields omitted
}

The internal redis response parser.

Methods

impl<'a, T: Read> Parser<T>

The parser can be used to parse redis responses into values. Generally you normally do not use this directly as it's already done for you by the client but in some more complex situations it might be useful to be able to parse the redis responses.

fn new(reader: T) -> Parser<T>

Creates a new parser that parses the data behind the reader. More than one value can be behind the reader in which case the parser can be invoked multiple times. In other words: the stream does not have to be terminated.

fn parse_value(&mut self) -> RedisResult<Value>

parses a single value out of the stream. If there are multiple values you can call this multiple times. If the reader is not yet ready this will block.