Struct redis::ScriptInvocationExperimental

pub struct ScriptInvocation<'a> {
    // some fields omitted
}

Represents a prepared script call.

Methods

impl<'a> ScriptInvocation<'a>

This type collects keys and other arguments for the script so that it can be then invoked. While the Script type itself holds the script, the ScriptInvocation holds the arguments that should be invoked until it's sent to the server.

fn arg<T: ToRedisArgs>(&'a mut self, arg: T) -> &'a mut ScriptInvocation

Adds a regular argument to the invocation. This ends up as ARGV[i] in the script.

fn key<T: ToRedisArgs>(&'a mut self, key: T) -> &'a mut ScriptInvocation

Adds a key argument to the invocation. This ends up as KEYS[i] in the script.

fn invoke<T: FromRedisValue>(&self, con: &ConnectionLike) -> RedisResult<T>

Invokes the script and returns the result.