redis

Trait ToSingleRedisArg

pub trait ToSingleRedisArg: ToRedisArgs { }

This trait marks that a value is serialized only into a single Redis value.

This should be implemented only for types that are serialized into exactly one value, otherwise the compiler can’t ensure the correctness of some commands.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.