Enum redis::acl::Rule

pub enum Rule {

    On,
    Off,
    AddCommand(String),
    RemoveCommand(String),
    AddCategory(String),
    RemoveCategory(String),
    AllCommands,
    NoCommands,
    AddPass(String),
    RemovePass(String),
    AddHashedPass(String),
    RemoveHashedPass(String),
    NoPass,
    ResetPass,
    Pattern(String),
    AllKeys,
    ResetKeys,
    Reset,

}

ACL rules are used in order to activate or remove a flag, or to perform a given change to the user ACL, which under the hood are just single words.

Variants

On

Enable the user: it is possible to authenticate as this user.

Off

Disable the user: it’s no longer possible to authenticate with this user, however the already authenticated connections will still work.

AddCommand(String)

Add the command to the list of commands the user can call.

RemoveCommand(String)

Remove the command to the list of commands the user can call.

AddCategory(String)

Add all the commands in such category to be called by the user.

RemoveCategory(String)

Remove the commands from such category the client can call.

AllCommands

Alias for +@all. Note that it implies the ability to execute all the future commands loaded via the modules system.

NoCommands

Alias for -@all.

AddPass(String)

Add this password to the list of valid password for the user.

RemovePass(String)

Remove this password from the list of valid passwords.

AddHashedPass(String)

Add this SHA-256 hash value to the list of valid passwords for the user.

RemoveHashedPass(String)

Remove this hash value from from the list of valid passwords

NoPass

All the set passwords of the user are removed, and the user is flagged as requiring no password: it means that every password will work against this user.

ResetPass

Flush the list of allowed passwords. Moreover removes the nopass status.

Pattern(String)

Add a pattern of keys that can be mentioned as part of commands.

AllKeys

Alias for ~*.

ResetKeys

Flush the list of allowed keys patterns.

Reset

Performs the following actions: resetpass, resetkeys, off, -@all. The user returns to the same state it has immediately after its creation.