Module redis::aio

Adds experimental async IO support to redis.

Modules

tokio

Enables the tokio compatibility

Structs

Connection

Represents a stateful redis TCP connection.

ConnectionManager

A ConnectionManager is a proxy that wraps a multiplexed connection and automatically reconnects to the server when necessary.

Monitor

Represents a Monitor connection.

MultiplexedConnection

A connection object which can be cloned, allowing requests to be be sent concurrently on the same underlying connection (tcp/unix socket).

PubSub

Represents a PubSub connection.

Traits

AsyncStream

Trait for objects that implements AsyncRead and AsyncWrite

ConnectionLike

An async abstraction over connections.