redis
Module aio

Adds async IO support to redis.

Modules

tokio

Enables the tokio compatibility

Structs

ConnectionDeprecated

Represents a stateful redis TCP connection.

Monitor

Represents a Monitor connection.

MultiplexedConnection

A connection object which can be cloned, allowing requests to be be sent concurrently on the same underlying connection (tcp/unix socket).

PubSub

A connection dedicated to pubsub messages.

PubSubSink

The sink part of a split async Pubsub.

PubSubStream

The stream part of a split async Pubsub.

SendError

An error showing that the receiver

Traits

AsyncPushSender

A trait for sender parts of a channel that can be used for sending push messages from async connection.

AsyncStream

Trait for objects that implements AsyncRead and AsyncWrite

ConnectionLike

An async abstraction over connections.