Module redis::aio

Adds async IO support to redis.

Modules

tokio

Enables the tokio compatibility

Structs

ConnectionDeprecated

Represents a stateful redis TCP connection.

Monitor

Represents a Monitor connection.

MultiplexedConnection

A connection object which can be cloned, allowing requests to be be sent concurrently on the same underlying connection (tcp/unix socket).

PubSub

Represents a PubSub connection.

Traits

AsyncStream

Trait for objects that implements AsyncRead and AsyncWrite

ConnectionLike

An async abstraction over connections.