redis::aio

Trait ConnectionLike

pub trait ConnectionLike {
    // Required methods
    fn req_packed_command<'a>(
        &'a mut self,
        cmd: &'a Cmd,
    ) -> RedisFuture<'a, Value>;
    fn get_db(&self) -> i64;
}

An async abstraction over connections.

Required Methods

fn req_packed_command<'a>(&'a mut self, cmd: &'a Cmd) -> RedisFuture<'a, Value>

Sends an already encoded (packed) command into the TCP socket and reads the single response from it.

fn get_db(&self) -> i64

Returns the database this connection is bound to. Note that this information might be unreliable because it’s initially cached and also might be incorrect if the connection like object is not actually connected.