Function redis::parse_redis_valueExperimental

pub fn parse_redis_value(bytes: &[u8]) -> RedisResult<Value>

Parses bytes into a redis value.

This is the most straightforward way to parse something into a low level redis value instead of having to use a whole parser.