Struct redis::geo::RadiusOptions

pub struct RadiusOptions { /* fields omitted */ }

Options for the GEORADIUS and GEORADIUSBYMEMBER commands

Example

use redis::{Commands, RedisResult};
use redis::geo::{RadiusSearchResult, RadiusOptions, RadiusOrder, Unit};
fn nearest_in_radius(
    con: &mut redis::Connection,
    key: &str,
    longitude: f64,
    latitude: f64,
    meters: f64,
    limit: usize,
) -> RedisResult<Vec<RadiusSearchResult>> {
    let opts = RadiusOptions::default()
        .order(RadiusOrder::Asc)
        .limit(limit);
    con.geo_radius(key, longitude, latitude, meters, Unit::Meters, opts)
}

Implementations

impl RadiusOptions

pub fn limit(self, n: usize) -> Self

Limit the results to the first N matching items.

pub fn with_dist(self) -> Self

Return the distance of the returned items from the specified center. The distance is returned in the same unit as the unit specified as the radius argument of the command.

pub fn with_coord(self) -> Self

Return the longitude, latitude coordinates of the matching items.

pub fn order(self, o: RadiusOrder) -> Self

Sort the returned items

pub fn store<K: ToRedisArgs>(self, key: K) -> Self

Store the results in a sorted set at key, instead of returning them.

This feature can’t be used with any with_* method.

pub fn store_dist<K: ToRedisArgs>(self, key: K) -> Self

Store the results in a sorted set at key, with the distance from the center as its score. This feature can’t be used with any with_* method.