redis

Module streams

Defines types to use with the streams commands.

Structs

StreamAddOptions

Builder options for xadd_options command

StreamAutoClaimOptions

Builder options for xautoclaim_options command.

StreamAutoClaimReply

Reply type used with the xautoclaim_options command.

StreamClaimOptions

Builder options for xclaim_options command.

StreamClaimReply

Reply type used with xclaim command.

StreamId

Represents a stream id and its field/values as a HashMap

StreamInfoConsumer

A consumer parsed from xinfo_consumers command.

StreamInfoConsumersReply

Reply type used with xinfo_consumer command, an array of every consumer in a specific consumer group.

StreamInfoGroup

A group parsed from xinfo_groups command.

StreamInfoGroupsReply

Reply type used with xinfo_groups command.

StreamInfoStreamReply

Reply type used with xinfo_stream command, containing general information about the stream stored at the specified key.

StreamKey

Represents a stream key and its id’s parsed from xread methods.

StreamPendingCountReply

Reply type used with xpending_count and xpending_consumer_count commands.

StreamPendingData

Inner reply type when an xpending command has data.

StreamPendingId

Represents a pending message parsed from xpending methods.

StreamRangeReply

Reply type used with xrange, xrange_count, xrange_all, xrevrange, xrevrange_count, xrevrange_all commands.

StreamReadOptions

Builder options for xread_options command.

StreamReadReply

Reply type used with xread or xread_options commands.

StreamTrimOptions

Builder options for xtrim_options command

Enums

StreamDeletionPolicy

Deletion policy for stream entries.

StreamMaxlen

Utility enum for passing MAXLEN [= or ~] [COUNT] arguments into StreamCommands. The enum value represents the count.

StreamPendingReply

Reply type used with xpending command.

StreamTrimStrategy

Utility enum for passing <MAXLEN|MINID> [=|~] threshold [LIMIT count] arguments into StreamCommands. The enum values the trimming mode (=|~), the threshold, and the optional limit

StreamTrimmingMode

Utility enum for passing the trim mode[=|~] arguments into StreamCommands.

XAckDelStatusCode

Status codes returned by the XACKDEL command

XDelExStatusCode

Status codes returned by the XDELEX command