Struct redis::Client

pub struct Client { /* private fields */ }

The client type.

Implementations

impl Client

The client acts as connector to the redis server. By itself it does not do much other than providing a convenient way to fetch a connection from it. In the future the plan is to provide a connection pool in the client.

When opening a client a URL in the following format should be used:

redis://host:port/db

Example usage::

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();

pub fn open<T: IntoConnectionInfo>(params: T) -> RedisResult<Client>

Connects to a redis server and returns a client. This does not actually open a connection yet but it does perform some basic checks on the URL that might make the operation fail.

pub fn get_connection(&self) -> RedisResult<Connection>

Instructs the client to actually connect to redis and returns a connection object. The connection object can be used to send commands to the server. This can fail with a variety of errors (like unreachable host) so it’s important that you handle those errors.

pub fn get_connection_with_timeout(
&self,
timeout: Duration
) -> RedisResult<Connection>

Instructs the client to actually connect to redis with specified timeout and returns a connection object. The connection object can be used to send commands to the server. This can fail with a variety of errors (like unreachable host) so it’s important that you handle those errors.

pub fn get_connection_info(&self) -> &ConnectionInfo

Returns a reference of client connection info object.

impl Client

To enable async support you need to chose one of the supported runtimes and active its corresponding feature: tokio-comp or async-std-comp

pub async fn get_async_connection(&self) -> RedisResult<Connection>

Returns an async connection from the client.

pub async fn get_tokio_connection(&self) -> RedisResult<Connection>

Returns an async connection from the client.

pub async fn get_multiplexed_async_connection(
&self
) -> RedisResult<MultiplexedConnection>

Returns an async connection from the client.

pub async fn get_multiplexed_tokio_connection(
&self
) -> RedisResult<MultiplexedConnection>

Returns an async multiplexed connection from the client.

A multiplexed connection can be cloned, allowing requests to be be sent concurrently on the same underlying connection (tcp/unix socket).

pub async fn create_multiplexed_tokio_connection(
&self
) -> RedisResult<(MultiplexedConnection, impl Future<Output = ()>)>

Returns an async multiplexed connection from the client and a future which must be polled to drive any requests submitted to it (see get_multiplexed_tokio_connection).

A multiplexed connection can be cloned, allowing requests to be be sent concurrently on the same underlying connection (tcp/unix socket).

pub async fn get_tokio_connection_manager(
&self
) -> RedisResult<ConnectionManager>

Returns an async ConnectionManager from the client.

The connection manager wraps a MultiplexedConnection. If a command to that connection fails with a connection error, then a new connection is established in the background and the error is returned to the caller.

This means that on connection loss at least one command will fail, but the connection will be re-established automatically if possible. Please refer to the ConnectionManager docs for detailed reconnecting behavior.

A connection manager can be cloned, allowing requests to be be sent concurrently on the same underlying connection (tcp/unix socket).