Struct redis::ClientExperimental

pub struct Client {
    // some fields omitted
}

The client type.

Methods

impl Client

The client acts as connector to the redis server. By itself it does not do much other than providing a convenient way to fetch a connection from it. In the future the plan is to provide a connection pool in the client.

When opening a client a URL in the following format should be used:

redis://host:port/db

Example usage::

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();

fn open(uri: &str) -> IoResult<Client>

Connects to a redis server and returns a client. This does not actually open a connection yet but it does perform some basic checks on the URL that might make the operation fail.

fn get_connection(&self) -> IoResult<Connection>

Instructs the client to actually connect to redis and returns a connection object. The connection object can be used to send commands to the server. This can fail with a variety of errors (like unreachable host) so it's important that you handle those errors.

fn get_pubsub(&self) -> IoResult<PubSub>

Returns a PubSub connection. A pubsub connection can be used to listen to messages coming in through the redis publish/subscribe system.

Note that redis' pubsub operates across all databases.