Struct redis::Cmd

pub struct Cmd { /* private fields */ }

Represents redis commands.

Implementations

impl Cmd

A command acts as a builder interface to creating encoded redis requests. This allows you to easiy assemble a packed command by chaining arguments together.

Basic example:

redis::Cmd::new().arg("SET").arg("my_key").arg(42);

There is also a helper function called cmd which makes it a tiny bit shorter:

redis::cmd("SET").arg("my_key").arg(42);

Because currently rust’s currently does not have an ideal system for lifetimes of temporaries, sometimes you need to hold on to the initially generated command:

let mut cmd = redis::cmd("SMEMBERS");
let mut iter : redis::Iter<i32> = cmd.arg("my_set").clone().iter(&mut con).unwrap();

pub fn new() -> Cmd

Creates a new empty command.

pub fn arg<T: ToRedisArgs>(&mut self, arg: T) -> &mut Cmd

Appends an argument to the command. The argument passed must be a type that implements ToRedisArgs. Most primitive types as well as vectors of primitive types implement it.

For instance all of the following are valid:

redis::cmd("SET").arg(&["my_key", "my_value"]);
redis::cmd("SET").arg("my_key").arg(42);
redis::cmd("SET").arg("my_key").arg(b"my_value");

pub fn cursor_arg(&mut self, cursor: u64) -> &mut Cmd

Works similar to arg but adds a cursor argument. This is always an integer and also flips the command implementation to support a different mode for the iterators where the iterator will ask for another batch of items when the local data is exhausted.

let mut cmd = redis::cmd("SSCAN");
let mut iter : redis::Iter<isize> =
    cmd.arg("my_set").cursor_arg(0).clone().iter(&mut con).unwrap();
for x in iter {
    // do something with the item
}

pub fn get_packed_command(&self) -> Vec<u8>

Returns the packed command as a byte vector.

pub fn in_scan_mode(&self) -> bool

Returns true if the command is in scan mode.

pub fn query<T: FromRedisValue>(
&self,
con: &mut dyn ConnectionLike
) -> RedisResult<T>

Sends the command as query to the connection and converts the result to the target redis value. This is the general way how you can retrieve data.

pub async fn query_async<C, T: FromRedisValue>(
&self,
con: &mut C
) -> RedisResult<T> where
C: ConnectionLike,

Async version of query.

pub fn iter<T: FromRedisValue>(
self,
con: &mut dyn ConnectionLike
) -> RedisResult<Iter<'_, T>>

Similar to query() but returns an iterator over the items of the bulk result or iterator. In normal mode this is not in any way more efficient than just querying into a Vec<T> as it’s internally implemented as buffering into a vector. This however is useful when cursor_arg was used in which case the iterator will query for more items until the server side cursor is exhausted.

This is useful for commands such as SSCAN, SCAN and others.

One speciality of this function is that it will check if the response looks like a cursor or not and always just looks at the payload. This way you can use the function the same for responses in the format of KEYS (just a list) as well as SSCAN (which returns a tuple of cursor and list).

pub async fn iter_async<'a, T: FromRedisValue + 'a>(
self,
con: &'a mut (dyn AsyncConnection + Send)
) -> RedisResult<AsyncIter<'a, T>>

Similar to iter() but returns an AsyncIter over the items of the bulk result or iterator. A futures::Stream can be obtained by calling stream() on the AsyncIter. In normal mode this is not in any way more efficient than just querying into a Vec<T> as it’s internally implemented as buffering into a vector. This however is useful when cursor_arg was used in which case the stream will query for more items until the server side cursor is exhausted.

This is useful for commands such as SSCAN, SCAN and others in async contexts.

One speciality of this function is that it will check if the response looks like a cursor or not and always just looks at the payload. This way you can use the function the same for responses in the format of KEYS (just a list) as well as SSCAN (which returns a tuple of cursor and list).

pub fn execute(&self, con: &mut dyn ConnectionLike)

This is a shortcut to query() that does not return a value and will fail the task if the query fails because of an error. This is mainly useful in examples and for simple commands like setting keys.

This is equivalent to a call of query like this:

let _ : () = redis::cmd("PING").query(&mut con).unwrap();

pub fn args_iter(
&self
) -> impl Iterator<Item = Arg<&[u8]>> + Clone + ExactSizeIterator

Returns an iterator over the arguments in this command (including the command name itself)

impl Cmd

pub fn get<'a, K: ToRedisArgs>(key: K) -> Self

Get the value of a key. If key is a vec this becomes an MGET.

pub fn keys<'a, K: ToRedisArgs>(key: K) -> Self

Gets all keys matching pattern

pub fn set<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) -> Self

Set the string value of a key.

pub fn set_multiple<'a, K: ToRedisArgs, V: ToRedisArgs>(
items: &'a [(K, V)]
) -> Self

Sets multiple keys to their values.

pub fn set_ex<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
value: V,
seconds: usize
) -> Self

Set the value and expiration of a key.

pub fn pset_ex<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
value: V,
milliseconds: usize
) -> Self

Set the value and expiration in milliseconds of a key.

pub fn set_nx<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) -> Self

Set the value of a key, only if the key does not exist

pub fn mset_nx<'a, K: ToRedisArgs, V: ToRedisArgs>(items: &'a [(K, V)]) -> Self

Sets multiple keys to their values failing if at least one already exists.

pub fn getset<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) -> Self

Set the string value of a key and return its old value.

pub fn getrange<'a, K: ToRedisArgs>(key: K, from: isize, to: isize) -> Self

Get a range of bytes/substring from the value of a key. Negative values provide an offset from the end of the value.

pub fn setrange<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
offset: isize,
value: V
) -> Self

Overwrite the part of the value stored in key at the specified offset.

pub fn del<'a, K: ToRedisArgs>(key: K) -> Self

Delete one or more keys.

pub fn exists<'a, K: ToRedisArgs>(key: K) -> Self

Determine if a key exists.

pub fn expire<'a, K: ToRedisArgs>(key: K, seconds: usize) -> Self

Set a key’s time to live in seconds.

pub fn expire_at<'a, K: ToRedisArgs>(key: K, ts: usize) -> Self

Set the expiration for a key as a UNIX timestamp.

pub fn pexpire<'a, K: ToRedisArgs>(key: K, ms: usize) -> Self

Set a key’s time to live in milliseconds.

pub fn pexpire_at<'a, K: ToRedisArgs>(key: K, ts: usize) -> Self

Set the expiration for a key as a UNIX timestamp in milliseconds.

pub fn persist<'a, K: ToRedisArgs>(key: K) -> Self

Remove the expiration from a key.

pub fn ttl<'a, K: ToRedisArgs>(key: K) -> Self

Get the expiration time of a key.

pub fn pttl<'a, K: ToRedisArgs>(key: K) -> Self

Get the expiration time of a key in milliseconds.

pub fn rename<'a, K: ToRedisArgs>(key: K, new_key: K) -> Self

Rename a key.

pub fn rename_nx<'a, K: ToRedisArgs>(key: K, new_key: K) -> Self

Rename a key, only if the new key does not exist.

pub fn unlink<'a, K: ToRedisArgs>(key: K) -> Self

Unlink one or more keys.

pub fn append<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) -> Self

Append a value to a key.

pub fn incr<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, delta: V) -> Self

Increment the numeric value of a key by the given amount. This issues a INCRBY or INCRBYFLOAT depending on the type.

pub fn decr<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, delta: V) -> Self

Decrement the numeric value of a key by the given amount.

pub fn setbit<'a, K: ToRedisArgs>(key: K, offset: usize, value: bool) -> Self

Sets or clears the bit at offset in the string value stored at key.

pub fn getbit<'a, K: ToRedisArgs>(key: K, offset: usize) -> Self

Returns the bit value at offset in the string value stored at key.

pub fn bitcount<'a, K: ToRedisArgs>(key: K) -> Self

Count set bits in a string.

pub fn bitcount_range<'a, K: ToRedisArgs>(
key: K,
start: usize,
end: usize
) -> Self

Count set bits in a string in a range.

pub fn bit_and<'a, K: ToRedisArgs>(dstkey: K, srckeys: K) -> Self

Perform a bitwise AND between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_or<'a, K: ToRedisArgs>(dstkey: K, srckeys: K) -> Self

Perform a bitwise OR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_xor<'a, K: ToRedisArgs>(dstkey: K, srckeys: K) -> Self

Perform a bitwise XOR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_not<'a, K: ToRedisArgs>(dstkey: K, srckey: K) -> Self

Perform a bitwise NOT of the key (containing string values) and store the result in the destination key.

pub fn strlen<'a, K: ToRedisArgs>(key: K) -> Self

Get the length of the value stored in a key.

pub fn hget<'a, K: ToRedisArgs, F: ToRedisArgs>(key: K, field: F) -> Self

Gets a single (or multiple) fields from a hash.

pub fn hdel<'a, K: ToRedisArgs, F: ToRedisArgs>(key: K, field: F) -> Self

Deletes a single (or multiple) fields from a hash.

pub fn hset<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
key: K,
field: F,
value: V
) -> Self

Sets a single field in a hash.

pub fn hset_nx<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
key: K,
field: F,
value: V
) -> Self

Sets a single field in a hash if it does not exist.

pub fn hset_multiple<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
key: K,
items: &'a [(F, V)]
) -> Self

Sets a multiple fields in a hash.

pub fn hincr<'a, K: ToRedisArgs, F: ToRedisArgs, D: ToRedisArgs>(
key: K,
field: F,
delta: D
) -> Self

Increments a value.

pub fn hexists<'a, K: ToRedisArgs, F: ToRedisArgs>(key: K, field: F) -> Self

Checks if a field in a hash exists.

pub fn hkeys<'a, K: ToRedisArgs>(key: K) -> Self

Gets all the keys in a hash.

pub fn hvals<'a, K: ToRedisArgs>(key: K) -> Self

Gets all the values in a hash.

pub fn hgetall<'a, K: ToRedisArgs>(key: K) -> Self

Gets all the fields and values in a hash.

pub fn hlen<'a, K: ToRedisArgs>(key: K) -> Self

Gets the length of a hash.

pub fn blpop<'a, K: ToRedisArgs>(key: K, timeout: usize) -> Self

Remove and get the first element in a list, or block until one is available.

pub fn brpop<'a, K: ToRedisArgs>(key: K, timeout: usize) -> Self

Remove and get the last element in a list, or block until one is available.

pub fn brpoplpush<'a, K: ToRedisArgs>(
srckey: K,
dstkey: K,
timeout: usize
) -> Self

Pop a value from a list, push it to another list and return it; or block until one is available.

pub fn lindex<'a, K: ToRedisArgs>(key: K, index: isize) -> Self

Get an element from a list by its index.

pub fn linsert_before<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs>(
key: K,
pivot: P,
value: V
) -> Self

Insert an element before another element in a list.

pub fn linsert_after<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs>(
key: K,
pivot: P,
value: V
) -> Self

Insert an element after another element in a list.

pub fn llen<'a, K: ToRedisArgs>(key: K) -> Self

Returns the length of the list stored at key.

pub fn lpop<'a, K: ToRedisArgs>(key: K) -> Self

Removes and returns the first element of the list stored at key.

pub fn lpos<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
value: V,
options: LposOptions
) -> Self

Returns the index of the first matching value of the list stored at key.

pub fn lpush<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) -> Self

Insert all the specified values at the head of the list stored at key.

pub fn lpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
value: V
) -> Self

Inserts a value at the head of the list stored at key, only if key already exists and holds a list.

pub fn lrange<'a, K: ToRedisArgs>(key: K, start: isize, stop: isize) -> Self

Returns the specified elements of the list stored at key.

pub fn lrem<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
count: isize,
value: V
) -> Self

Removes the first count occurrences of elements equal to value from the list stored at key.

pub fn ltrim<'a, K: ToRedisArgs>(key: K, start: isize, stop: isize) -> Self

Trim an existing list so that it will contain only the specified range of elements specified.

pub fn lset<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
index: isize,
value: V
) -> Self

Sets the list element at index to value

pub fn rpop<'a, K: ToRedisArgs>(key: K) -> Self

Removes and returns the last element of the list stored at key.

pub fn rpoplpush<'a, K: ToRedisArgs>(key: K, dstkey: K) -> Self

Pop a value from a list, push it to another list and return it.

pub fn rpush<'a, K: ToRedisArgs, V: ToRedisArgs>(key: K, value: V) -> Self

Insert all the specified values at the tail of the list stored at key.

pub fn rpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs>(
key: K,
value: V
) -> Self

Inserts value at the tail of the list stored at key, only if key already exists and holds a list.

pub fn sadd<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, member: M) -> Self

Add one or more members to a set.

pub fn scard<'a, K: ToRedisArgs>(key: K) -> Self

Get the number of members in a set.

pub fn sdiff<'a, K: ToRedisArgs>(keys: K) -> Self

Subtract multiple sets.

pub fn sdiffstore<'a, K: ToRedisArgs>(dstkey: K, keys: K) -> Self

Subtract multiple sets and store the resulting set in a key.

pub fn sinter<'a, K: ToRedisArgs>(keys: K) -> Self

Intersect multiple sets.

pub fn sinterstore<'a, K: ToRedisArgs>(dstkey: K, keys: K) -> Self

Intersect multiple sets and store the resulting set in a key.

pub fn sismember<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, member: M) -> Self

Determine if a given value is a member of a set.

pub fn smembers<'a, K: ToRedisArgs>(key: K) -> Self

Get all the members in a set.

pub fn smove<'a, K: ToRedisArgs, M: ToRedisArgs>(
srckey: K,
dstkey: K,
member: M
) -> Self

Move a member from one set to another.

pub fn spop<'a, K: ToRedisArgs>(key: K) -> Self

Remove and return a random member from a set.

pub fn srandmember<'a, K: ToRedisArgs>(key: K) -> Self

Get one random member from a set.

pub fn srandmember_multiple<'a, K: ToRedisArgs>(key: K, count: usize) -> Self

Get multiple random members from a set.

pub fn srem<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, member: M) -> Self

Remove one or more members from a set.

pub fn sunion<'a, K: ToRedisArgs>(keys: K) -> Self

Add multiple sets.

pub fn sunionstore<'a, K: ToRedisArgs>(dstkey: K, keys: K) -> Self

Add multiple sets and store the resulting set in a key.

pub fn zadd<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(
key: K,
member: M,
score: S
) -> Self

Add one member to a sorted set, or update its score if it already exists.

pub fn zadd_multiple<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(
key: K,
items: &'a [(S, M)]
) -> Self

Add multiple members to a sorted set, or update its score if it already exists.

pub fn zcard<'a, K: ToRedisArgs>(key: K) -> Self

Get the number of members in a sorted set.

pub fn zcount<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM
) -> Self

Count the members in a sorted set with scores within the given values.

pub fn zincr<'a, K: ToRedisArgs, M: ToRedisArgs, D: ToRedisArgs>(
key: K,
member: M,
delta: D
) -> Self

Increments the member in a sorted set at key by delta. If the member does not exist, it is added with delta as its score.

pub fn zinterstore<'a, K: ToRedisArgs>(dstkey: K, keys: &'a [K]) -> Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zinterstore_min<'a, K: ToRedisArgs>(dstkey: K, keys: &'a [K]) -> Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zinterstore_max<'a, K: ToRedisArgs>(dstkey: K, keys: &'a [K]) -> Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn zlexcount<'a, K: ToRedisArgs, L: ToRedisArgs>(
key: K,
min: L,
max: L
) -> Self

Count the number of members in a sorted set between a given lexicographical range.

pub fn zpopmax<'a, K: ToRedisArgs>(key: K, count: isize) -> Self

Removes and returns up to count members with the highest scores in a sorted set

pub fn zpopmin<'a, K: ToRedisArgs>(key: K, count: isize) -> Self

Removes and returns up to count members with the lowest scores in a sorted set

pub fn zrange<'a, K: ToRedisArgs>(key: K, start: isize, stop: isize) -> Self

Return a range of members in a sorted set, by index

pub fn zrange_withscores<'a, K: ToRedisArgs>(
key: K,
start: isize,
stop: isize
) -> Self

Return a range of members in a sorted set, by index with scores.

pub fn zrangebylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM
) -> Self

Return a range of members in a sorted set, by lexicographical range.

pub fn zrangebylex_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> Self

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrevrangebylex<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
key: K,
max: MM,
min: M
) -> Self

Return a range of members in a sorted set, by lexicographical range.

pub fn zrevrangebylex_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> Self

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrangebyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM
) -> Self

Return a range of members in a sorted set, by score.

pub fn zrangebyscore_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM
) -> Self

Return a range of members in a sorted set, by score with scores.

pub fn zrangebyscore_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> Self

Return a range of members in a sorted set, by score with limit.

pub fn zrangebyscore_limit_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> Self

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrank<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, member: M) -> Self

Determine the index of a member in a sorted set.

pub fn zrem<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) -> Self

Remove one or more members from a sorted set.

pub fn zrembylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM
) -> Self

Remove all members in a sorted set between the given lexicographical range.

pub fn zremrangebyrank<'a, K: ToRedisArgs>(
key: K,
start: isize,
stop: isize
) -> Self

Remove all members in a sorted set within the given indexes.

pub fn zrembyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
key: K,
min: M,
max: MM
) -> Self

Remove all members in a sorted set within the given scores.

pub fn zrevrange<'a, K: ToRedisArgs>(key: K, start: isize, stop: isize) -> Self

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrange_withscores<'a, K: ToRedisArgs>(
key: K,
start: isize,
stop: isize
) -> Self

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrangebyscore<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
key: K,
max: MM,
min: M
) -> Self

Return a range of members in a sorted set, by score.

pub fn zrevrangebyscore_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
key: K,
max: MM,
min: M
) -> Self

Return a range of members in a sorted set, by score with scores.

pub fn zrevrangebyscore_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> Self

Return a range of members in a sorted set, by score with limit.

pub fn zrevrangebyscore_limit_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> Self

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrevrank<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, member: M) -> Self

Determine the index of a member in a sorted set, with scores ordered from high to low.

pub fn zscore<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, member: M) -> Self

Get the score associated with the given member in a sorted set.

pub fn zunionstore<'a, K: ToRedisArgs>(dstkey: K, keys: &'a [K]) -> Self

Unions multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zunionstore_min<'a, K: ToRedisArgs>(dstkey: K, keys: &'a [K]) -> Self

Unions multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zunionstore_max<'a, K: ToRedisArgs>(dstkey: K, keys: &'a [K]) -> Self

Unions multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn pfadd<'a, K: ToRedisArgs, E: ToRedisArgs>(key: K, element: E) -> Self

Adds the specified elements to the specified HyperLogLog.

pub fn pfcount<'a, K: ToRedisArgs>(key: K) -> Self

Return the approximated cardinality of the set(s) observed by the HyperLogLog at key(s).

pub fn pfmerge<'a, K: ToRedisArgs>(dstkey: K, srckeys: K) -> Self

Merge N different HyperLogLogs into a single one.

pub fn publish<'a, K: ToRedisArgs, E: ToRedisArgs>(
channel: K,
message: E
) -> Self

Posts a message to the given channel.

pub fn acl_load<'a>() -> Self

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will reload the ACLs from the file, replacing all the current ACL rules with the ones defined in the file.

pub fn acl_save<'a>() -> Self

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will save the currently defined ACLs from the server memory to the ACL file.

pub fn acl_list<'a>() -> Self

Shows the currently active ACL rules in the Redis server.

pub fn acl_users<'a>() -> Self

Shows a list of all the usernames of the currently configured users in the Redis ACL system.

pub fn acl_getuser<'a, K: ToRedisArgs>(username: K) -> Self

Returns all the rules defined for an existing ACL user.

pub fn acl_setuser<'a, K: ToRedisArgs>(username: K) -> Self

Creates an ACL user without any privilege.

pub fn acl_setuser_rules<'a, K: ToRedisArgs>(
username: K,
rules: &'a [Rule]
) -> Self

Creates an ACL user with the specified rules or modify the rules of an existing user.

pub fn acl_deluser<'a, K: ToRedisArgs>(usernames: &'a [K]) -> Self

Delete all the specified ACL users and terminate all the connections that are authenticated with such users.

pub fn acl_cat<'a>() -> Self

Shows the available ACL categories.

pub fn acl_cat_categoryname<'a, K: ToRedisArgs>(categoryname: K) -> Self

Shows all the Redis commands in the specified category.

pub fn acl_genpass<'a>() -> Self

Generates a 256-bits password starting from /dev/urandom if available.

pub fn acl_genpass_bits<'a>(bits: isize) -> Self

Generates a 1-to-1024-bits password starting from /dev/urandom if available.

pub fn acl_whoami<'a>() -> Self

Returns the username the current connection is authenticated with.

pub fn acl_log<'a>(count: isize) -> Self

Shows a list of recent ACL security events

pub fn acl_log_reset<'a>() -> Self

Clears the ACL log.

pub fn acl_help<'a>() -> Self

Returns a helpful text describing the different subcommands.

pub fn geo_add<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) -> Self

Adds the specified geospatial items to the specified key.

Every member has to be written as a tuple of (longitude, latitude, member_name). It can be a single tuple, or a vector of tuples.

longitude, latitude can be set using redis::geo::Coord.

Returns the number of elements added to the sorted set, not including elements already existing for which the score was updated.

Example

use redis::{Commands, Connection, RedisResult};
use redis::geo::Coord;

fn add_point(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", (Coord::lon_lat(13.361389, 38.115556), "Palermo"))
}

fn add_point_with_tuples(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", ("13.361389", "38.115556", "Palermo"))
}

fn add_many_points(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", &[
        ("13.361389", "38.115556", "Palermo"),
        ("15.087269", "37.502669", "Catania")
    ])
}

pub fn geo_dist<'a, K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs>(
key: K,
member1: M1,
member2: M2,
unit: Unit
) -> Self

Return the distance between two members in the geospatial index represented by the sorted set.

If one or both the members are missing, the command returns NULL, so it may be convenient to parse its response as either Option<f64> or Option<String>.

Example

use redis::{Commands, RedisResult};
use redis::geo::Unit;

fn get_dists(con: &mut redis::Connection) {
    let x: RedisResult<f64> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Catania",
        Unit::Kilometers
    );
    // x is Ok(166.2742)

    let x: RedisResult<Option<f64>> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Atlantis",
        Unit::Meters
    );
    // x is Ok(None)
}

pub fn geo_hash<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) -> Self

Return valid Geohash strings representing the position of one or more members of the geospatial index represented by the sorted set at key.

Example

use redis::{Commands, RedisResult};

fn get_hash(con: &mut redis::Connection) {
    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", "Palermo");
    // x is vec!["sqc8b49rny0"]

    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", &["Palermo", "Catania"]);
    // x is vec!["sqc8b49rny0", "sqdtr74hyu0"]
}

pub fn geo_pos<'a, K: ToRedisArgs, M: ToRedisArgs>(key: K, members: M) -> Self

Return the positions of all the specified members of the geospatial index represented by the sorted set at key.

Every position is a pair of (longitude, latitude). redis::geo::Coord can be used to convert these value in a struct.

Example

use redis::{Commands, RedisResult};
use redis::geo::Coord;

fn get_position(con: &mut redis::Connection) {
    let x: RedisResult<Vec<Vec<f64>>> = con.geo_pos("my_gis", &["Palermo", "Catania"]);
    // x is [ [ 13.361389, 38.115556 ], [ 15.087269, 37.502669 ] ];

    let x: Vec<Coord<f64>> = con.geo_pos("my_gis", "Palermo").unwrap();
    // x[0].longitude is 13.361389
    // x[0].latitude is 38.115556
}

pub fn geo_radius<'a, K: ToRedisArgs>(
key: K,
longitude: f64,
latitude: f64,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> Self

Return the members of a sorted set populated with geospatial information using geo_add, which are within the borders of the area specified with the center location and the maximum distance from the center (the radius).

Every item in the result can be read with redis::geo::RadiusSearchResult, which support the multiple formats returned by GEORADIUS.

use redis::{Commands, RedisResult};
use redis::geo::{RadiusOptions, RadiusSearchResult, RadiusOrder, Unit};

fn radius(con: &mut redis::Connection) -> Vec<RadiusSearchResult> {
    let opts = RadiusOptions::default().with_dist().order(RadiusOrder::Asc);
    con.geo_radius("my_gis", 15.90, 37.21, 51.39, Unit::Kilometers, opts).unwrap()
}

pub fn geo_radius_by_member<'a, K: ToRedisArgs, M: ToRedisArgs>(
key: K,
member: M,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> Self

Retrieve members selected by distance with the center of member. The member itself is always contained in the results.

pub fn xack<'a, K: ToRedisArgs, G: ToRedisArgs, I: ToRedisArgs>(
key: K,
group: G,
ids: &'a [I]
) -> Self

Ack pending stream messages checked out by a consumer.

XACK <key> <group> <id> <id> ... <id>

pub fn xadd<'a, K: ToRedisArgs, ID: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
key: K,
id: ID,
items: &'a [(F, V)]
) -> Self

Add a stream message by key. Use * as the id for the current timestamp.

XADD key <ID or *> [field value] [field value] ...

pub fn xadd_map<'a, K: ToRedisArgs, ID: ToRedisArgs, BTM: ToRedisArgs>(
key: K,
id: ID,
map: BTM
) -> Self

BTreeMap variant for adding a stream message by key. Use * as the id for the current timestamp.

XADD key <ID or *> [rust BTreeMap] ...

pub fn xadd_maxlen<'a, K: ToRedisArgs, ID: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
key: K,
maxlen: StreamMaxlen,
id: ID,
items: &'a [(F, V)]
) -> Self

Add a stream message while capping the stream at a maxlength.

XADD key [MAXLEN [~|=] <count>] <ID or *> [field value] [field value] ...

pub fn xadd_maxlen_map<'a, K: ToRedisArgs, ID: ToRedisArgs, BTM: ToRedisArgs>(
key: K,
maxlen: StreamMaxlen,
id: ID,
map: BTM
) -> Self

BTreeMap variant for adding a stream message while capping the stream at a maxlength.

XADD key [MAXLEN [~|=] <count>] <ID or *> [rust BTreeMap] ...

pub fn xclaim<'a, K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs, MIT: ToRedisArgs, ID: ToRedisArgs>(
key: K,
group: G,
consumer: C,
min_idle_time: MIT,
ids: &'a [ID]
) -> Self

Claim pending, unacked messages, after some period of time, currently checked out by another consumer.

This method only accepts the must-have arguments for claiming messages. If optional arguments are required, see xclaim_options below.

XCLAIM <key> <group> <consumer> <min-idle-time> [<ID-1> <ID-2>]

pub fn xclaim_options<'a, K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs, MIT: ToRedisArgs, ID: ToRedisArgs>(
key: K,
group: G,
consumer: C,
min_idle_time: MIT,
ids: &'a [ID],
options: StreamClaimOptions
) -> Self

This is the optional arguments version for claiming unacked, pending messages currently checked out by another consumer.

use redis::{Connection,Commands,RedisResult};
use redis::streams::{StreamClaimOptions,StreamClaimReply};
let client = redis::Client::open("redis://127.0.0.1/0").unwrap();
let mut con = client.get_connection().unwrap();

// Claim all pending messages for key "k1",
// from group "g1", checked out by consumer "c1"
// for 10ms with RETRYCOUNT 2 and FORCE

let opts = StreamClaimOptions::default()
    .with_force()
    .retry(2);
let results: RedisResult<StreamClaimReply> =
    con.xclaim_options("k1", "g1", "c1", 10, &["0"], opts);

// All optional arguments return a `Result<StreamClaimReply>` with one exception:
// Passing JUSTID returns only the message `id` and omits the HashMap for each message.

let opts = StreamClaimOptions::default()
    .with_justid();
let results: RedisResult<Vec<String>> =
    con.xclaim_options("k1", "g1", "c1", 10, &["0"], opts);

XCLAIM <key> <group> <consumer> <min-idle-time> <ID-1> <ID-2>
    [IDLE <milliseconds>] [TIME <mstime>] [RETRYCOUNT <count>]
    [FORCE] [JUSTID]

pub fn xdel<'a, K: ToRedisArgs, ID: ToRedisArgs>(key: K, ids: &'a [ID]) -> Self

Deletes a list of ids for a given stream key.

XDEL <key> [<ID1> <ID2> ... <IDN>]

pub fn xgroup_create<'a, K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs>(
key: K,
group: G,
id: ID
) -> Self

This command is used for creating a consumer group. It expects the stream key to already exist. Otherwise, use xgroup_create_mkstream if it doesn’t. The id is the starting message id all consumers should read from. Use $ If you want all consumers to read from the last message added to stream.

XGROUP CREATE <key> <groupname> <id or $>

pub fn xgroup_create_mkstream<'a, K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs>(
key: K,
group: G,
id: ID
) -> Self

This is the alternate version for creating a consumer group which makes the stream if it doesn’t exist.

XGROUP CREATE <key> <groupname> <id or $> [MKSTREAM]

pub fn xgroup_setid<'a, K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs>(
key: K,
group: G,
id: ID
) -> Self

Alter which id you want consumers to begin reading from an existing consumer group.

XGROUP SETID <key> <groupname> <id or $>

pub fn xgroup_destroy<'a, K: ToRedisArgs, G: ToRedisArgs>(
key: K,
group: G
) -> Self

Destroy an existing consumer group for a given stream key

XGROUP SETID <key> <groupname> <id or $>

pub fn xgroup_delconsumer<'a, K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs>(
key: K,
group: G,
consumer: C
) -> Self

This deletes a consumer from an existing consumer group for given stream `key.

XGROUP DELCONSUMER <key> <groupname> <consumername>

pub fn xinfo_consumers<'a, K: ToRedisArgs, G: ToRedisArgs>(
key: K,
group: G
) -> Self

This returns all info details about which consumers have read messages for given consumer group. Take note of the StreamInfoConsumersReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO CONSUMERS <key> <group>

pub fn xinfo_groups<'a, K: ToRedisArgs>(key: K) -> Self

Returns all consumer groups created for a given stream key. Take note of the StreamInfoGroupsReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO GROUPS <key>

pub fn xinfo_stream<'a, K: ToRedisArgs>(key: K) -> Self

Returns info about high-level stream details (first & last message id, length, number of groups, etc.) Take note of the StreamInfoStreamReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO STREAM <key>

pub fn xlen<'a, K: ToRedisArgs>(key: K) -> Self

Returns the number of messages for a given stream key.

XLEN <key>

pub fn xpending<'a, K: ToRedisArgs, G: ToRedisArgs>(key: K, group: G) -> Self

This is a basic version of making XPENDING command calls which only passes a stream key and consumer group and it returns details about which consumers have pending messages that haven’t been acked.

You can use this method along with xclaim or xclaim_options for determining which messages need to be retried.

Take note of the StreamPendingReply return type.

XPENDING <key> <group> [<start> <stop> <count> [<consumer>]]

pub fn xpending_count<'a, K: ToRedisArgs, G: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs>(
key: K,
group: G,
start: S,
end: E,
count: C
) -> Self

This XPENDING version returns a list of all messages over the range. You can use this for paginating pending messages (but without the message HashMap).

Start and end follow the same rules xrange args. Set start to - and end to + for the entire stream.

Take note of the StreamPendingCountReply return type.

XPENDING <key> <group> <start> <stop> <count>

pub fn xpending_consumer_count<'a, K: ToRedisArgs, G: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs, CN: ToRedisArgs>(
key: K,
group: G,
start: S,
end: E,
count: C,
consumer: CN
) -> Self

An alternate version of xpending_count which filters by consumer name.

Start and end follow the same rules xrange args. Set start to - and end to + for the entire stream.

Take note of the StreamPendingCountReply return type.

XPENDING <key> <group> <start> <stop> <count> <consumer>

pub fn xrange<'a, K: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs>(
key: K,
start: S,
end: E
) -> Self

Returns a range of messages in a given stream key.

Set start to - to begin at the first message. Set end to + to end the most recent message. You can pass message id to both start and end.

Take note of the StreamRangeReply return type.

XRANGE key start end

pub fn xrange_all<'a, K: ToRedisArgs>(key: K) -> Self

A helper method for automatically returning all messages in a stream by key. Use with caution!

XRANGE key - +

pub fn xrange_count<'a, K: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs>(
key: K,
start: S,
end: E,
count: C
) -> Self

A method for paginating a stream by key.

XRANGE key start end [COUNT <n>]

pub fn xread<'a, K: ToRedisArgs, ID: ToRedisArgs>(
keys: &'a [K],
ids: &'a [ID]
) -> Self

Read a list of ids for each stream key. This is the basic form of reading streams. For more advanced control, like blocking, limiting, or reading by consumer group, see xread_options.

XREAD STREAMS key_1 key_2 ... key_N ID_1 ID_2 ... ID_N

pub fn xread_options<'a, K: ToRedisArgs, ID: ToRedisArgs>(
keys: &'a [K],
ids: &'a [ID],
options: StreamReadOptions
) -> Self

This method handles setting optional arguments for XREAD or XREADGROUP Redis commands.

use redis::{Connection,RedisResult,Commands};
use redis::streams::{StreamReadOptions,StreamReadReply};
let client = redis::Client::open("redis://127.0.0.1/0").unwrap();
let mut con = client.get_connection().unwrap();

// Read 10 messages from the start of the stream,
// without registering as a consumer group.

let opts = StreamReadOptions::default()
    .count(10);
let results: RedisResult<StreamReadReply> =
    con.xread_options(&["k1"], &["0"], opts);

// Read all undelivered messages for a given
// consumer group. Be advised: the consumer group must already
// exist before making this call. Also note: we're passing
// '>' as the id here, which means all undelivered messages.

let opts = StreamReadOptions::default()
    .group("group-1", "consumer-1");
let results: RedisResult<StreamReadReply> =
    con.xread_options(&["k1"], &[">"], opts);

XREAD [BLOCK <milliseconds>] [COUNT <count>]
    STREAMS key_1 key_2 ... key_N
    ID_1 ID_2 ... ID_N

XREADGROUP [BLOCK <milliseconds>] [COUNT <count>] [NOACK] [GROUP group-name consumer-name]
    STREAMS key_1 key_2 ... key_N
    ID_1 ID_2 ... ID_N

pub fn xrevrange<'a, K: ToRedisArgs, E: ToRedisArgs, S: ToRedisArgs>(
key: K,
end: E,
start: S
) -> Self

This is the reverse version of xrange. The same rules apply for start and end here.

XREVRANGE key end start

pub fn xrevrange_all<'a, K: ToRedisArgs>(key: K) -> Self

This is the reverse version of xrange_all. The same rules apply for start and end here.

XREVRANGE key + -

pub fn xrevrange_count<'a, K: ToRedisArgs, E: ToRedisArgs, S: ToRedisArgs, C: ToRedisArgs>(
key: K,
end: E,
start: S,
count: C
) -> Self

This is the reverse version of xrange_count. The same rules apply for start and end here.

XREVRANGE key end start [COUNT <n>]

pub fn xtrim<'a, K: ToRedisArgs>(key: K, maxlen: StreamMaxlen) -> Self

Trim a stream key to a MAXLEN count.

XTRIM <key> MAXLEN [~|=] <count>  (Same as XADD MAXLEN option)