Struct redis::Connection

pub struct Connection {
    // some fields omitted
}

Represents a stateful redis TCP connection.

Methods

impl Connection

A connection is an object that represents a single redis connection. It provides basic support for sending encoded commands into a redis connection and to read a response from it. It's bound to a single database and can only be created from the client.

You generally do not much with this object other than passing it to Cmd objects.

fn send_packed_command(&self, cmd: &[u8]) -> RedisResult<()>

Sends an already encoded (packed) command into the TCP socket and does not read a response. This is useful for commands like MONITOR which yield multiple items. This needs to be used with care because it changes the state of the connection.

fn recv_response(&self) -> RedisResult<Value>

Fetches a single response from the connection. This is useful if used in combination with send_packed_command.