redis

Struct InfoDict

pub struct InfoDict { /* private fields */ }

An info dictionary type.

Implementations

impl InfoDict

This type provides convenient access to key/value data returned by the “INFO” command. It acts like a regular mapping but also has a convenience method get which can return data in the appropriate type.

For instance this can be used to query the server for the role it’s in (master, slave) etc:

let info : redis::InfoDict = redis::cmd("INFO").query(&mut con)?;
let role : Option<String> = info.get("role");

pub fn new(kvpairs: &str) -> InfoDict

Creates a new info dictionary from a string in the response of the INFO command. Each line is a key, value pair with the key and value separated by a colon (:). Lines starting with a hash (#) are ignored.

pub fn get<T: FromRedisValue>(&self, key: &str) -> Option<T>

Fetches a value by key and converts it into the given type. Typical types are String, bool and integer types.

pub fn find(&self, key: &&str) -> Option<&Value>

Looks up a key in the info dict.

pub fn contains_key(&self, key: &&str) -> bool

Checks if a key is contained in the info dicf.

pub fn len(&self) -> usize

Returns the size of the info dict.

pub fn is_empty(&self) -> bool

Checks if the dict is empty.