Struct redis::Pipeline

pub struct Pipeline { /* fields omitted */ }

Represents a redis command pipeline.

Methods

impl Pipeline

A pipeline allows you to send multiple commands in one go to the redis server. API wise it's very similar to just using a command but it allows multiple commands to be chained and some features such as iteration are not available.

Basic example:

let ((k1, k2),) : ((i32, i32),) = redis::pipe()
    .cmd("SET").arg("key_1").arg(42).ignore()
    .cmd("SET").arg("key_2").arg(43).ignore()
    .cmd("MGET").arg(&["key_1", "key_2"]).query(&mut con).unwrap();

As you can see with cmd you can start a new command. By default each command produces a value but for some you can ignore them by calling ignore on the command. That way it will be skipped in the return value which is useful for SET commands and others, which do not have a useful return value.

pub fn new() -> Pipeline

Creates an empty pipeline. For consistency with the cmd api a pipe function is provided as alias.

pub fn with_capacity(capacity: usize) -> Pipeline

Creates an empty pipeline with pre-allocated capacity.

pub fn cmd(&mut self, name: &str) -> &mut Pipeline

Starts a new command. Functions such as arg then become available to add more arguments to that command.

pub fn add_command(&mut self, cmd: Cmd) -> &mut Pipeline

Adds a command to the pipeline.

pub fn arg<T: ToRedisArgs>(&mut self, arg: T) -> &mut Pipeline

Adds an argument to the last started command. This works similar to the arg method of the Cmd object.

Note that this function fails the task if executed on an empty pipeline.

pub fn ignore(&mut self) -> &mut Pipeline

Instructs the pipeline to ignore the return value of this command. It will still be ensured that it is not an error, but any successful result is just thrown away. This makes result processing through tuples much easier because you do not need to handle all the items you do not care about.

Note that this function fails the task if executed on an empty pipeline.

pub fn atomic(&mut self) -> &mut Pipeline

This enables atomic mode. In atomic mode the whole pipeline is enclosed in MULTI/EXEC. From the user's point of view nothing changes however. This is easier than using MULTI/EXEC yourself as the format does not change.

let (k1, k2) : (i32, i32) = redis::pipe()
    .atomic()
    .cmd("GET").arg("key_1")
    .cmd("GET").arg("key_2").query(&mut con).unwrap();

pub fn cmd_iter(&self) -> impl Iterator<Item = &Cmd>

Returns an iterator over all the commands currently in this pipeline

pub fn get_packed_pipeline(&self) -> Vec<u8>

Returns the encoded pipeline commands.

pub fn query<T: FromRedisValue>(
&self,
con: &mut dyn ConnectionLike
) -> RedisResult<T>

Executes the pipeline and fetches the return values. Since most pipelines return different types it's recommended to use tuple matching to process the results:

let (k1, k2) : (i32, i32) = redis::pipe()
    .cmd("SET").arg("key_1").arg(42).ignore()
    .cmd("SET").arg("key_2").arg(43).ignore()
    .cmd("GET").arg("key_1")
    .cmd("GET").arg("key_2").query(&mut con).unwrap();

NOTE: A Pipeline object may be reused after query() with all the commands as were inserted to them. In order to clear a Pipeline object with minimal memory released/allocated, it is necessary to call the clear() before inserting new commands.

pub fn clear(&mut self)

Clear a Pipeline object internal data structure.

This allows reusing a Pipeline object as a clear object while performing a minimal amount of memory released/reallocated.

pub async fn query_async<'_, '_, C, T: FromRedisValue>(
&'_ self,
con: &'_ mut C
) -> RedisResult<T> where
C: ConnectionLike,

Async version of query.

pub fn execute(&self, con: &mut dyn ConnectionLike)

This is a shortcut to query() that does not return a value and will fail the task if the query of the pipeline fails.

This is equivalent to a call of query like this:

let _ : () = redis::pipe().cmd("PING").query(&mut con).unwrap();

NOTE: A Pipeline object may be reused after query() with all the commands as were inserted to them. In order to clear a Pipeline object with minimal memory released/allocated, it is necessary to call the clear() before inserting new commands.

impl Pipeline

Implements common redis commands for pipelines. Unlike the regular commands trait, this returns the pipeline rather than a result directly. Other than that it works the same however.

pub fn get<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the value of a key. If key is a vec this becomes an MGET.

pub fn keys<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all keys matching pattern

pub fn set<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Set the string value of a key.

pub fn set_multiple<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
items: &'a [(K, V)]
) -> &mut Self

Sets multiple keys to their values.

pub fn set_ex<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V,
seconds: usize
) -> &mut Self

Set the value and expiration of a key.

pub fn pset_ex<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V,
milliseconds: usize
) -> &mut Self

Set the value and expiration in milliseconds of a key.

pub fn set_nx<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Set the value of a key, only if the key does not exist

pub fn mset_nx<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
items: &'a [(K, V)]
) -> &mut Self

Sets multiple keys to their values failing if at least one already exists.

pub fn getset<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Set the string value of a key and return its old value.

pub fn getrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
from: isize,
to: isize
) -> &mut Self

Get a range of bytes/substring from the value of a key. Negative values provide an offset from the end of the value.

pub fn setrange<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
offset: isize,
value: V
) -> &mut Self

Overwrite the part of the value stored in key at the specified offset.

pub fn del<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Delete one or more keys.

pub fn exists<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Determine if a key exists.

pub fn expire<'a, K: ToRedisArgs>(
&mut self,
key: K,
seconds: usize
) -> &mut Self

Set a key's time to live in seconds.

pub fn expire_at<'a, K: ToRedisArgs>(&mut self, key: K, ts: usize) -> &mut Self

Set the expiration for a key as a UNIX timestamp.

pub fn pexpire<'a, K: ToRedisArgs>(&mut self, key: K, ms: usize) -> &mut Self

Set a key's time to live in milliseconds.

pub fn pexpire_at<'a, K: ToRedisArgs>(&mut self, key: K, ts: usize) -> &mut Self

Set the expiration for a key as a UNIX timestamp in milliseconds.

pub fn persist<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Remove the expiration from a key.

pub fn ttl<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the expiration time of a key.

pub fn pttl<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the expiration time of a key in milliseconds.

pub fn rename<'a, K: ToRedisArgs>(&mut self, key: K, new_key: K) -> &mut Self

Rename a key.

pub fn rename_nx<'a, K: ToRedisArgs>(&mut self, key: K, new_key: K) -> &mut Self

Rename a key, only if the new key does not exist.

pub fn append<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Append a value to a key.

pub fn incr<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
delta: V
) -> &mut Self

Increment the numeric value of a key by the given amount. This issues a INCRBY or INCRBYFLOAT depending on the type.

pub fn setbit<'a, K: ToRedisArgs>(
&mut self,
key: K,
offset: usize,
value: bool
) -> &mut Self

Sets or clears the bit at offset in the string value stored at key.

pub fn getbit<'a, K: ToRedisArgs>(&mut self, key: K, offset: usize) -> &mut Self

Returns the bit value at offset in the string value stored at key.

pub fn bitcount<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Count set bits in a string.

pub fn bitcount_range<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: usize,
end: usize
) -> &mut Self

Count set bits in a string in a range.

pub fn bit_and<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
srckeys: K
) -> &mut Self

Perform a bitwise AND between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_or<'a, K: ToRedisArgs>(&mut self, dstkey: K, srckeys: K) -> &mut Self

Perform a bitwise OR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_xor<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
srckeys: K
) -> &mut Self

Perform a bitwise XOR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_not<'a, K: ToRedisArgs>(&mut self, dstkey: K, srckey: K) -> &mut Self

Perform a bitwise NOT of the key (containing string values) and store the result in the destination key.

pub fn strlen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the length of the value stored in a key.

pub fn hget<'a, K: ToRedisArgs, F: ToRedisArgs>(
&mut self,
key: K,
field: F
) -> &mut Self

Gets a single (or multiple) fields from a hash.

pub fn hdel<'a, K: ToRedisArgs, F: ToRedisArgs>(
&mut self,
key: K,
field: F
) -> &mut Self

Deletes a single (or multiple) fields from a hash.

pub fn hset<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
field: F,
value: V
) -> &mut Self

Sets a single field in a hash.

pub fn hset_nx<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
field: F,
value: V
) -> &mut Self

Sets a single field in a hash if it does not exist.

pub fn hset_multiple<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
items: &'a [(F, V)]
) -> &mut Self

Sets a multiple fields in a hash.

pub fn hincr<'a, K: ToRedisArgs, F: ToRedisArgs, D: ToRedisArgs>(
&mut self,
key: K,
field: F,
delta: D
) -> &mut Self

Increments a value.

pub fn hexists<'a, K: ToRedisArgs, F: ToRedisArgs>(
&mut self,
key: K,
field: F
) -> &mut Self

Checks if a field in a hash exists.

pub fn hkeys<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all the keys in a hash.

pub fn hvals<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all the values in a hash.

pub fn hgetall<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all the fields and values in a hash.

pub fn hlen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets the length of a hash.

pub fn blpop<'a, K: ToRedisArgs>(&mut self, key: K, timeout: usize) -> &mut Self

Remove and get the first element in a list, or block until one is available.

pub fn brpop<'a, K: ToRedisArgs>(&mut self, key: K, timeout: usize) -> &mut Self

Remove and get the last element in a list, or block until one is available.

pub fn brpoplpush<'a, K: ToRedisArgs>(
&mut self,
srckey: K,
dstkey: K,
timeout: usize
) -> &mut Self

Pop a value from a list, push it to another list and return it; or block until one is available.

pub fn lindex<'a, K: ToRedisArgs>(&mut self, key: K, index: isize) -> &mut Self

Get an element from a list by its index.

pub fn linsert_before<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
pivot: P,
value: V
) -> &mut Self

Insert an element before another element in a list.

pub fn linsert_after<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
pivot: P,
value: V
) -> &mut Self

Insert an element after another element in a list.

pub fn llen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Returns the length of the list stored at key.

pub fn lpop<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Removes and returns the first element of the list stored at key.

pub fn lpush<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Insert all the specified values at the head of the list stored at key.

pub fn lpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Inserts a value at the head of the list stored at key, only if key already exists and holds a list.

pub fn lrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Returns the specified elements of the list stored at key.

pub fn lrem<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
count: isize,
value: V
) -> &mut Self

Removes the first count occurrences of elements equal to value from the list stored at key.

pub fn ltrim<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Trim an existing list so that it will contain only the specified range of elements specified.

pub fn lset<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
index: isize,
value: V
) -> &mut Self

Sets the list element at index to value

pub fn rpop<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Removes and returns the last element of the list stored at key.

pub fn rpoplpush<'a, K: ToRedisArgs>(&mut self, key: K, dstkey: K) -> &mut Self

Pop a value from a list, push it to another list and return it.

pub fn rpush<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Insert all the specified values at the tail of the list stored at key.

pub fn rpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Inserts value at the tail of the list stored at key, only if key already exists and holds a list.

pub fn sadd<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Add one or more members to a set.

pub fn scard<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the number of members in a set.

pub fn sdiff<'a, K: ToRedisArgs>(&mut self, keys: K) -> &mut Self

Subtract multiple sets.

pub fn sdiffstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: K
) -> &mut Self

Subtract multiple sets and store the resulting set in a key.

pub fn sinter<'a, K: ToRedisArgs>(&mut self, keys: K) -> &mut Self

Intersect multiple sets.

pub fn sinterstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: K
) -> &mut Self

Intersect multiple sets and store the resulting set in a key.

pub fn sismember<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Determine if a given value is a member of a set.

pub fn smembers<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get all the members in a set.

pub fn smove<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
srckey: K,
dstkey: K,
member: M
) -> &mut Self

Move a member from one set to another.

pub fn spop<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Remove and return a random member from a set.

pub fn srandmember<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get one random member from a set.

pub fn srandmember_multiple<'a, K: ToRedisArgs>(
&mut self,
key: K,
count: usize
) -> &mut Self

Get multiple random members from a set.

pub fn srem<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Remove one or more members from a set.

pub fn sunion<'a, K: ToRedisArgs>(&mut self, keys: K) -> &mut Self

Add multiple sets.

pub fn sunionstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: K
) -> &mut Self

Add multiple sets and store the resulting set in a key.

pub fn zadd<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M,
score: S
) -> &mut Self

Add one member to a sorted set, or update its score if it already exists.

pub fn zadd_multiple<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
items: &'a [(S, M)]
) -> &mut Self

Add multiple members to a sorted set, or update its score if it already exists.

pub fn zcard<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the number of members in a sorted set.

pub fn zcount<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Count the members in a sorted set with scores within the given values.

pub fn zincr<'a, K: ToRedisArgs, M: ToRedisArgs, D: ToRedisArgs>(
&mut self,
key: K,
member: M,
delta: D
) -> &mut Self

Increments the member in a sorted set at key by delta. If the member does not exist, it is added with delta as its score.

pub fn zinterstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zinterstore_min<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zinterstore_max<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn zlexcount<'a, K: ToRedisArgs, L: ToRedisArgs>(
&mut self,
key: K,
min: L,
max: L
) -> &mut Self

Count the number of members in a sorted set between a given lexicographical range.

pub fn zrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index

pub fn zrange_withscores<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index with scores.

pub fn zrangebylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range.

pub fn zrangebylex_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrevrangebylex<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range.

pub fn zrevrangebylex_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrangebyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Return a range of members in a sorted set, by score.

pub fn zrangebyscore_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Return a range of members in a sorted set, by score with scores.

pub fn zrangebyscore_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit.

pub fn zrangebyscore_limit_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrank<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Determine the index of a member in a sorted set.

pub fn zrem<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Remove one or more members from a sorted set.

pub fn zrembylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Remove all members in a sorted set between the given lexicographical range.

pub fn zremrangebyrank<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Remove all members in a sorted set within the given indexes.

pub fn zrembyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Remove all members in a sorted set within the given scores.

pub fn zrevrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrange_withscores<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrangebyscore<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M
) -> &mut Self

Return a range of members in a sorted set, by score.

pub fn zrevrangebyscore_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M
) -> &mut Self

Return a range of members in a sorted set, by score with scores.

pub fn zrevrangebyscore_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit.

pub fn zrevrangebyscore_limit_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrevrank<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Determine the index of a member in a sorted set, with scores ordered from high to low.

pub fn zscore<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Get the score associated with the given member in a sorted set.

pub fn zunionstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Unions multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zunionstore_min<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Unions multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zunionstore_max<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Unions multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn pfadd<'a, K: ToRedisArgs, E: ToRedisArgs>(
&mut self,
key: K,
element: E
) -> &mut Self

Adds the specified elements to the specified HyperLogLog.

pub fn pfcount<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Return the approximated cardinality of the set(s) observed by the HyperLogLog at key(s).

pub fn pfmerge<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
srckeys: K
) -> &mut Self

Merge N different HyperLogLogs into a single one.

pub fn publish<'a, K: ToRedisArgs, E: ToRedisArgs>(
&mut self,
channel: K,
message: E
) -> &mut Self

Posts a message to the given channel.

pub fn geo_add<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Adds the specified geospatial items to the specified key.

Every member has to be written as a tuple of (longitude, latitude, member_name). It can be a single tuple, or a vector of tuples.

longitude, latitude can be set using redis::geo::Coord.

Returns the number of elements added to the sorted set, not including elements already existing for which the score was updated.

Example

use redis::{Commands, Connection, RedisResult};
use redis::geo::Coord;

fn add_point(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", (Coord::lon_lat(13.361389, 38.115556), "Palermo"))
}

fn add_point_with_tuples(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", ("13.361389", "38.115556", "Palermo"))
}

fn add_many_points(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", &[
        ("13.361389", "38.115556", "Palermo"),
        ("15.087269", "37.502669", "Catania")
    ])
}

pub fn geo_dist<'a, K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs>(
&mut self,
key: K,
member1: M1,
member2: M2,
unit: Unit
) -> &mut Self

Return the distance between two members in the geospatial index represented by the sorted set.

If one or both the members are missing, the command returns NULL, so it may be convenient to parse its response as either Option<f64> or Option<String>.

Example

use redis::{Commands, RedisResult};
use redis::geo::Unit;

fn get_dists(con: &mut redis::Connection) {
    let x: RedisResult<f64> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Catania",
        Unit::Kilometers
    );
    // x is Ok(166.2742)

    let x: RedisResult<Option<f64>> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Atlantis",
        Unit::Meters
    );
    // x is Ok(None)
}

pub fn geo_hash<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Return valid Geohash strings representing the position of one or more members of the geospatial index represented by the sorted set at key.

Example

use redis::{Commands, RedisResult};

fn get_hash(con: &mut redis::Connection) {
    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", "Palermo");
    // x is vec!["sqc8b49rny0"]

    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", &["Palermo", "Catania"]);
    // x is vec!["sqc8b49rny0", "sqdtr74hyu0"]
}

pub fn geo_pos<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Return the positions of all the specified members of the geospatial index represented by the sorted set at key.

Every position is a pair of (longitude, latitude). redis::geo::Coord can be used to convert these value in a struct.

Example

use redis::{Commands, RedisResult};
use redis::geo::Coord;

fn get_position(con: &mut redis::Connection) {
    let x: RedisResult<Vec<Vec<f64>>> = con.geo_pos("my_gis", &["Palermo", "Catania"]);
    // x is [ [ 13.361389, 38.115556 ], [ 15.087269, 37.502669 ] ];

    let x: Vec<Coord<f64>> = con.geo_pos("my_gis", "Palermo").unwrap();
    // x[0].longitude is 13.361389
    // x[0].latitude is 38.115556
}

pub fn geo_radius<'a, K: ToRedisArgs>(
&mut self,
key: K,
longitude: f64,
latitude: f64,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> &mut Self

Return the members of a sorted set populated with geospatial information using geo_add, which are within the borders of the area specified with the center location and the maximum distance from the center (the radius).

Every item in the result can be read with redis::geo::RadiusSearchResult, which support the multiple formats returned by GEORADIUS.

use redis::{Commands, RedisResult};
use redis::geo::{RadiusOptions, RadiusSearchResult, RadiusOrder, Unit};

fn radius(con: &mut redis::Connection) -> Vec<RadiusSearchResult> {
    let opts = RadiusOptions::default().with_dist().order(RadiusOrder::Asc);
    con.geo_radius("my_gis", 15.90, 37.21, 51.39, Unit::Kilometers, opts).unwrap()
}

pub fn geo_radius_by_member<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> &mut Self

Retrieve members selected by distance with the center of member. The member itself is always contained in the results.