Struct redis::Pipeline

pub struct Pipeline { /* fields omitted */ }

Represents a redis command pipeline.

Implementations

impl Pipeline

A pipeline allows you to send multiple commands in one go to the redis server. API wise it’s very similar to just using a command but it allows multiple commands to be chained and some features such as iteration are not available.

Basic example:

let ((k1, k2),) : ((i32, i32),) = redis::pipe()
    .cmd("SET").arg("key_1").arg(42).ignore()
    .cmd("SET").arg("key_2").arg(43).ignore()
    .cmd("MGET").arg(&["key_1", "key_2"]).query(&mut con).unwrap();

As you can see with cmd you can start a new command. By default each command produces a value but for some you can ignore them by calling ignore on the command. That way it will be skipped in the return value which is useful for SET commands and others, which do not have a useful return value.

pub fn new() -> Pipeline

Creates an empty pipeline. For consistency with the cmd api a pipe function is provided as alias.

pub fn with_capacity(capacity: usize) -> Pipeline

Creates an empty pipeline with pre-allocated capacity.

pub fn atomic(&mut self) -> &mut Pipeline

This enables atomic mode. In atomic mode the whole pipeline is enclosed in MULTI/EXEC. From the user’s point of view nothing changes however. This is easier than using MULTI/EXEC yourself as the format does not change.

let (k1, k2) : (i32, i32) = redis::pipe()
    .atomic()
    .cmd("GET").arg("key_1")
    .cmd("GET").arg("key_2").query(&mut con).unwrap();

pub fn get_packed_pipeline(&self) -> Vec<u8>

Returns the encoded pipeline commands.

pub fn query<T: FromRedisValue>(
&self,
con: &mut dyn ConnectionLike
) -> RedisResult<T>

Executes the pipeline and fetches the return values. Since most pipelines return different types it’s recommended to use tuple matching to process the results:

let (k1, k2) : (i32, i32) = redis::pipe()
    .cmd("SET").arg("key_1").arg(42).ignore()
    .cmd("SET").arg("key_2").arg(43).ignore()
    .cmd("GET").arg("key_1")
    .cmd("GET").arg("key_2").query(&mut con).unwrap();

NOTE: A Pipeline object may be reused after query() with all the commands as were inserted to them. In order to clear a Pipeline object with minimal memory released/allocated, it is necessary to call the clear() before inserting new commands.

pub async fn query_async<C, T: FromRedisValue>(
&self,
con: &mut C
) -> RedisResult<T> where
C: ConnectionLike,

Async version of query.

pub fn execute(&self, con: &mut dyn ConnectionLike)

This is a shortcut to query() that does not return a value and will fail the task if the query of the pipeline fails.

This is equivalent to a call of query like this:

let _ : () = redis::pipe().cmd("PING").query(&mut con).unwrap();

NOTE: A Pipeline object may be reused after query() with all the commands as were inserted to them. In order to clear a Pipeline object with minimal memory released/allocated, it is necessary to call the clear() before inserting new commands.

impl Pipeline

pub fn add_command(&mut self, cmd: Cmd) -> &mut Self

Adds a command to the cluster pipeline.

pub fn cmd(&mut self, name: &str) -> &mut Self

Starts a new command. Functions such as arg then become available to add more arguments to that command.

pub fn cmd_iter(&self) -> impl Iterator<Item = &Cmd>

Returns an iterator over all the commands currently in this pipeline

pub fn ignore(&mut self) -> &mut Self

Instructs the pipeline to ignore the return value of this command. It will still be ensured that it is not an error, but any successful result is just thrown away. This makes result processing through tuples much easier because you do not need to handle all the items you do not care about.

pub fn arg<T: ToRedisArgs>(&mut self, arg: T) -> &mut Self

Adds an argument to the last started command. This works similar to the arg method of the Cmd object.

Note that this function fails the task if executed on an empty pipeline.

pub fn clear(&mut self)

Clear a pipeline object’s internal data structure.

This allows reusing a pipeline object as a clear object while performing a minimal amount of memory released/reallocated.

impl Pipeline

Implements common redis commands for pipelines. Unlike the regular commands trait, this returns the pipeline rather than a result directly. Other than that it works the same however.

pub fn get<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the value of a key. If key is a vec this becomes an MGET.

pub fn keys<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all keys matching pattern

pub fn set<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Set the string value of a key.

pub fn set_multiple<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
items: &'a [(K, V)]
) -> &mut Self

Sets multiple keys to their values.

pub fn set_ex<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V,
seconds: usize
) -> &mut Self

Set the value and expiration of a key.

pub fn pset_ex<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V,
milliseconds: usize
) -> &mut Self

Set the value and expiration in milliseconds of a key.

pub fn set_nx<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Set the value of a key, only if the key does not exist

pub fn mset_nx<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
items: &'a [(K, V)]
) -> &mut Self

Sets multiple keys to their values failing if at least one already exists.

pub fn getset<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Set the string value of a key and return its old value.

pub fn getrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
from: isize,
to: isize
) -> &mut Self

Get a range of bytes/substring from the value of a key. Negative values provide an offset from the end of the value.

pub fn setrange<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
offset: isize,
value: V
) -> &mut Self

Overwrite the part of the value stored in key at the specified offset.

pub fn del<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Delete one or more keys.

pub fn exists<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Determine if a key exists.

pub fn expire<'a, K: ToRedisArgs>(
&mut self,
key: K,
seconds: usize
) -> &mut Self

Set a key’s time to live in seconds.

pub fn expire_at<'a, K: ToRedisArgs>(&mut self, key: K, ts: usize) -> &mut Self

Set the expiration for a key as a UNIX timestamp.

pub fn pexpire<'a, K: ToRedisArgs>(&mut self, key: K, ms: usize) -> &mut Self

Set a key’s time to live in milliseconds.

pub fn pexpire_at<'a, K: ToRedisArgs>(&mut self, key: K, ts: usize) -> &mut Self

Set the expiration for a key as a UNIX timestamp in milliseconds.

pub fn persist<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Remove the expiration from a key.

pub fn ttl<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the expiration time of a key.

pub fn pttl<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the expiration time of a key in milliseconds.

pub fn rename<'a, K: ToRedisArgs>(&mut self, key: K, new_key: K) -> &mut Self

Rename a key.

pub fn rename_nx<'a, K: ToRedisArgs>(&mut self, key: K, new_key: K) -> &mut Self

Rename a key, only if the new key does not exist.

pub fn unlink<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Unlink one or more keys.

pub fn append<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Append a value to a key.

pub fn incr<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
delta: V
) -> &mut Self

Increment the numeric value of a key by the given amount. This issues a INCRBY or INCRBYFLOAT depending on the type.

pub fn decr<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
delta: V
) -> &mut Self

Decrement the numeric value of a key by the given amount.

pub fn setbit<'a, K: ToRedisArgs>(
&mut self,
key: K,
offset: usize,
value: bool
) -> &mut Self

Sets or clears the bit at offset in the string value stored at key.

pub fn getbit<'a, K: ToRedisArgs>(&mut self, key: K, offset: usize) -> &mut Self

Returns the bit value at offset in the string value stored at key.

pub fn bitcount<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Count set bits in a string.

pub fn bitcount_range<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: usize,
end: usize
) -> &mut Self

Count set bits in a string in a range.

pub fn bit_and<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
srckeys: K
) -> &mut Self

Perform a bitwise AND between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_or<'a, K: ToRedisArgs>(&mut self, dstkey: K, srckeys: K) -> &mut Self

Perform a bitwise OR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_xor<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
srckeys: K
) -> &mut Self

Perform a bitwise XOR between multiple keys (containing string values) and store the result in the destination key.

pub fn bit_not<'a, K: ToRedisArgs>(&mut self, dstkey: K, srckey: K) -> &mut Self

Perform a bitwise NOT of the key (containing string values) and store the result in the destination key.

pub fn strlen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the length of the value stored in a key.

pub fn hget<'a, K: ToRedisArgs, F: ToRedisArgs>(
&mut self,
key: K,
field: F
) -> &mut Self

Gets a single (or multiple) fields from a hash.

pub fn hdel<'a, K: ToRedisArgs, F: ToRedisArgs>(
&mut self,
key: K,
field: F
) -> &mut Self

Deletes a single (or multiple) fields from a hash.

pub fn hset<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
field: F,
value: V
) -> &mut Self

Sets a single field in a hash.

pub fn hset_nx<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
field: F,
value: V
) -> &mut Self

Sets a single field in a hash if it does not exist.

pub fn hset_multiple<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
items: &'a [(F, V)]
) -> &mut Self

Sets a multiple fields in a hash.

pub fn hincr<'a, K: ToRedisArgs, F: ToRedisArgs, D: ToRedisArgs>(
&mut self,
key: K,
field: F,
delta: D
) -> &mut Self

Increments a value.

pub fn hexists<'a, K: ToRedisArgs, F: ToRedisArgs>(
&mut self,
key: K,
field: F
) -> &mut Self

Checks if a field in a hash exists.

pub fn hkeys<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all the keys in a hash.

pub fn hvals<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all the values in a hash.

pub fn hgetall<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets all the fields and values in a hash.

pub fn hlen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Gets the length of a hash.

pub fn blpop<'a, K: ToRedisArgs>(&mut self, key: K, timeout: usize) -> &mut Self

Remove and get the first element in a list, or block until one is available.

pub fn brpop<'a, K: ToRedisArgs>(&mut self, key: K, timeout: usize) -> &mut Self

Remove and get the last element in a list, or block until one is available.

pub fn brpoplpush<'a, K: ToRedisArgs>(
&mut self,
srckey: K,
dstkey: K,
timeout: usize
) -> &mut Self

Pop a value from a list, push it to another list and return it; or block until one is available.

pub fn lindex<'a, K: ToRedisArgs>(&mut self, key: K, index: isize) -> &mut Self

Get an element from a list by its index.

pub fn linsert_before<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
pivot: P,
value: V
) -> &mut Self

Insert an element before another element in a list.

pub fn linsert_after<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
pivot: P,
value: V
) -> &mut Self

Insert an element after another element in a list.

pub fn llen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Returns the length of the list stored at key.

pub fn lpop<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Removes and returns the first element of the list stored at key.

pub fn lpos<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V,
options: LposOptions
) -> &mut Self

Returns the index of the first matching value of the list stored at key.

pub fn lpush<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Insert all the specified values at the head of the list stored at key.

pub fn lpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Inserts a value at the head of the list stored at key, only if key already exists and holds a list.

pub fn lrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Returns the specified elements of the list stored at key.

pub fn lrem<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
count: isize,
value: V
) -> &mut Self

Removes the first count occurrences of elements equal to value from the list stored at key.

pub fn ltrim<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Trim an existing list so that it will contain only the specified range of elements specified.

pub fn lset<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
index: isize,
value: V
) -> &mut Self

Sets the list element at index to value

pub fn rpop<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Removes and returns the last element of the list stored at key.

pub fn rpoplpush<'a, K: ToRedisArgs>(&mut self, key: K, dstkey: K) -> &mut Self

Pop a value from a list, push it to another list and return it.

pub fn rpush<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Insert all the specified values at the tail of the list stored at key.

pub fn rpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
value: V
) -> &mut Self

Inserts value at the tail of the list stored at key, only if key already exists and holds a list.

pub fn sadd<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Add one or more members to a set.

pub fn scard<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the number of members in a set.

pub fn sdiff<'a, K: ToRedisArgs>(&mut self, keys: K) -> &mut Self

Subtract multiple sets.

pub fn sdiffstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: K
) -> &mut Self

Subtract multiple sets and store the resulting set in a key.

pub fn sinter<'a, K: ToRedisArgs>(&mut self, keys: K) -> &mut Self

Intersect multiple sets.

pub fn sinterstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: K
) -> &mut Self

Intersect multiple sets and store the resulting set in a key.

pub fn sismember<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Determine if a given value is a member of a set.

pub fn smembers<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get all the members in a set.

pub fn smove<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
srckey: K,
dstkey: K,
member: M
) -> &mut Self

Move a member from one set to another.

pub fn spop<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Remove and return a random member from a set.

pub fn srandmember<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get one random member from a set.

pub fn srandmember_multiple<'a, K: ToRedisArgs>(
&mut self,
key: K,
count: usize
) -> &mut Self

Get multiple random members from a set.

pub fn srem<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Remove one or more members from a set.

pub fn sunion<'a, K: ToRedisArgs>(&mut self, keys: K) -> &mut Self

Add multiple sets.

pub fn sunionstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: K
) -> &mut Self

Add multiple sets and store the resulting set in a key.

pub fn zadd<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M,
score: S
) -> &mut Self

Add one member to a sorted set, or update its score if it already exists.

pub fn zadd_multiple<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
items: &'a [(S, M)]
) -> &mut Self

Add multiple members to a sorted set, or update its score if it already exists.

pub fn zcard<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Get the number of members in a sorted set.

pub fn zcount<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Count the members in a sorted set with scores within the given values.

pub fn zincr<'a, K: ToRedisArgs, M: ToRedisArgs, D: ToRedisArgs>(
&mut self,
key: K,
member: M,
delta: D
) -> &mut Self

Increments the member in a sorted set at key by delta. If the member does not exist, it is added with delta as its score.

pub fn zinterstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zinterstore_min<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zinterstore_max<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Intersect multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn zlexcount<'a, K: ToRedisArgs, L: ToRedisArgs>(
&mut self,
key: K,
min: L,
max: L
) -> &mut Self

Count the number of members in a sorted set between a given lexicographical range.

pub fn zpopmax<'a, K: ToRedisArgs>(&mut self, key: K, count: isize) -> &mut Self

Removes and returns up to count members with the highest scores in a sorted set

pub fn zpopmin<'a, K: ToRedisArgs>(&mut self, key: K, count: isize) -> &mut Self

Removes and returns up to count members with the lowest scores in a sorted set

pub fn zrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index

pub fn zrange_withscores<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index with scores.

pub fn zrangebylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range.

pub fn zrangebylex_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrevrangebylex<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range.

pub fn zrevrangebylex_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by lexicographical range with offset and limit.

pub fn zrangebyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Return a range of members in a sorted set, by score.

pub fn zrangebyscore_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Return a range of members in a sorted set, by score with scores.

pub fn zrangebyscore_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit.

pub fn zrangebyscore_limit_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrank<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Determine the index of a member in a sorted set.

pub fn zrem<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Remove one or more members from a sorted set.

pub fn zrembylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Remove all members in a sorted set between the given lexicographical range.

pub fn zremrangebyrank<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Remove all members in a sorted set within the given indexes.

pub fn zrembyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs>(
&mut self,
key: K,
min: M,
max: MM
) -> &mut Self

Remove all members in a sorted set within the given scores.

pub fn zrevrange<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrange_withscores<'a, K: ToRedisArgs>(
&mut self,
key: K,
start: isize,
stop: isize
) -> &mut Self

Return a range of members in a sorted set, by index, with scores ordered from high to low.

pub fn zrevrangebyscore<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M
) -> &mut Self

Return a range of members in a sorted set, by score.

pub fn zrevrangebyscore_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M
) -> &mut Self

Return a range of members in a sorted set, by score with scores.

pub fn zrevrangebyscore_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit.

pub fn zrevrangebyscore_limit_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> &mut Self

Return a range of members in a sorted set, by score with limit with scores.

pub fn zrevrank<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Determine the index of a member in a sorted set, with scores ordered from high to low.

pub fn zscore<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M
) -> &mut Self

Get the score associated with the given member in a sorted set.

pub fn zunionstore<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Unions multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

pub fn zunionstore_min<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Unions multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

pub fn zunionstore_max<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> &mut Self

Unions multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

pub fn pfadd<'a, K: ToRedisArgs, E: ToRedisArgs>(
&mut self,
key: K,
element: E
) -> &mut Self

Adds the specified elements to the specified HyperLogLog.

pub fn pfcount<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Return the approximated cardinality of the set(s) observed by the HyperLogLog at key(s).

pub fn pfmerge<'a, K: ToRedisArgs>(
&mut self,
dstkey: K,
srckeys: K
) -> &mut Self

Merge N different HyperLogLogs into a single one.

pub fn publish<'a, K: ToRedisArgs, E: ToRedisArgs>(
&mut self,
channel: K,
message: E
) -> &mut Self

Posts a message to the given channel.

pub fn acl_load<'a>(&mut self) -> &mut Self

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will reload the ACLs from the file, replacing all the current ACL rules with the ones defined in the file.

pub fn acl_save<'a>(&mut self) -> &mut Self

When Redis is configured to use an ACL file (with the aclfile configuration option), this command will save the currently defined ACLs from the server memory to the ACL file.

pub fn acl_list<'a>(&mut self) -> &mut Self

Shows the currently active ACL rules in the Redis server.

pub fn acl_users<'a>(&mut self) -> &mut Self

Shows a list of all the usernames of the currently configured users in the Redis ACL system.

pub fn acl_getuser<'a, K: ToRedisArgs>(&mut self, username: K) -> &mut Self

Returns all the rules defined for an existing ACL user.

pub fn acl_setuser<'a, K: ToRedisArgs>(&mut self, username: K) -> &mut Self

Creates an ACL user without any privilege.

pub fn acl_setuser_rules<'a, K: ToRedisArgs>(
&mut self,
username: K,
rules: &'a [Rule]
) -> &mut Self

Creates an ACL user with the specified rules or modify the rules of an existing user.

pub fn acl_deluser<'a, K: ToRedisArgs>(
&mut self,
usernames: &'a [K]
) -> &mut Self

Delete all the specified ACL users and terminate all the connections that are authenticated with such users.

pub fn acl_cat<'a>(&mut self) -> &mut Self

Shows the available ACL categories.

pub fn acl_cat_categoryname<'a, K: ToRedisArgs>(
&mut self,
categoryname: K
) -> &mut Self

Shows all the Redis commands in the specified category.

pub fn acl_genpass<'a>(&mut self) -> &mut Self

Generates a 256-bits password starting from /dev/urandom if available.

pub fn acl_genpass_bits<'a>(&mut self, bits: isize) -> &mut Self

Generates a 1-to-1024-bits password starting from /dev/urandom if available.

pub fn acl_whoami<'a>(&mut self) -> &mut Self

Returns the username the current connection is authenticated with.

pub fn acl_log<'a>(&mut self, count: isize) -> &mut Self

Shows a list of recent ACL security events

pub fn acl_log_reset<'a>(&mut self) -> &mut Self

Clears the ACL log.

pub fn acl_help<'a>(&mut self) -> &mut Self

Returns a helpful text describing the different subcommands.

pub fn geo_add<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Adds the specified geospatial items to the specified key.

Every member has to be written as a tuple of (longitude, latitude, member_name). It can be a single tuple, or a vector of tuples.

longitude, latitude can be set using redis::geo::Coord.

Returns the number of elements added to the sorted set, not including elements already existing for which the score was updated.

Example

use redis::{Commands, Connection, RedisResult};
use redis::geo::Coord;

fn add_point(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", (Coord::lon_lat(13.361389, 38.115556), "Palermo"))
}

fn add_point_with_tuples(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", ("13.361389", "38.115556", "Palermo"))
}

fn add_many_points(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", &[
        ("13.361389", "38.115556", "Palermo"),
        ("15.087269", "37.502669", "Catania")
    ])
}

pub fn geo_dist<'a, K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs>(
&mut self,
key: K,
member1: M1,
member2: M2,
unit: Unit
) -> &mut Self

Return the distance between two members in the geospatial index represented by the sorted set.

If one or both the members are missing, the command returns NULL, so it may be convenient to parse its response as either Option<f64> or Option<String>.

Example

use redis::{Commands, RedisResult};
use redis::geo::Unit;

fn get_dists(con: &mut redis::Connection) {
    let x: RedisResult<f64> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Catania",
        Unit::Kilometers
    );
    // x is Ok(166.2742)

    let x: RedisResult<Option<f64>> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Atlantis",
        Unit::Meters
    );
    // x is Ok(None)
}

pub fn geo_hash<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Return valid Geohash strings representing the position of one or more members of the geospatial index represented by the sorted set at key.

Example

use redis::{Commands, RedisResult};

fn get_hash(con: &mut redis::Connection) {
    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", "Palermo");
    // x is vec!["sqc8b49rny0"]

    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", &["Palermo", "Catania"]);
    // x is vec!["sqc8b49rny0", "sqdtr74hyu0"]
}

pub fn geo_pos<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
members: M
) -> &mut Self

Return the positions of all the specified members of the geospatial index represented by the sorted set at key.

Every position is a pair of (longitude, latitude). redis::geo::Coord can be used to convert these value in a struct.

Example

use redis::{Commands, RedisResult};
use redis::geo::Coord;

fn get_position(con: &mut redis::Connection) {
    let x: RedisResult<Vec<Vec<f64>>> = con.geo_pos("my_gis", &["Palermo", "Catania"]);
    // x is [ [ 13.361389, 38.115556 ], [ 15.087269, 37.502669 ] ];

    let x: Vec<Coord<f64>> = con.geo_pos("my_gis", "Palermo").unwrap();
    // x[0].longitude is 13.361389
    // x[0].latitude is 38.115556
}

pub fn geo_radius<'a, K: ToRedisArgs>(
&mut self,
key: K,
longitude: f64,
latitude: f64,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> &mut Self

Return the members of a sorted set populated with geospatial information using geo_add, which are within the borders of the area specified with the center location and the maximum distance from the center (the radius).

Every item in the result can be read with redis::geo::RadiusSearchResult, which support the multiple formats returned by GEORADIUS.

use redis::{Commands, RedisResult};
use redis::geo::{RadiusOptions, RadiusSearchResult, RadiusOrder, Unit};

fn radius(con: &mut redis::Connection) -> Vec<RadiusSearchResult> {
    let opts = RadiusOptions::default().with_dist().order(RadiusOrder::Asc);
    con.geo_radius("my_gis", 15.90, 37.21, 51.39, Unit::Kilometers, opts).unwrap()
}

pub fn geo_radius_by_member<'a, K: ToRedisArgs, M: ToRedisArgs>(
&mut self,
key: K,
member: M,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> &mut Self

Retrieve members selected by distance with the center of member. The member itself is always contained in the results.

pub fn xack<'a, K: ToRedisArgs, G: ToRedisArgs, I: ToRedisArgs>(
&mut self,
key: K,
group: G,
ids: &'a [I]
) -> &mut Self

Ack pending stream messages checked out by a consumer.

XACK <key> <group> <id> <id> ... <id>

pub fn xadd<'a, K: ToRedisArgs, ID: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
id: ID,
items: &'a [(F, V)]
) -> &mut Self

Add a stream message by key. Use * as the id for the current timestamp.

XADD key <ID or *> [field value] [field value] ...

pub fn xadd_map<'a, K: ToRedisArgs, ID: ToRedisArgs, BTM: ToRedisArgs>(
&mut self,
key: K,
id: ID,
map: BTM
) -> &mut Self

BTreeMap variant for adding a stream message by key. Use * as the id for the current timestamp.

XADD key <ID or *> [rust BTreeMap] ...

pub fn xadd_maxlen<'a, K: ToRedisArgs, ID: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs>(
&mut self,
key: K,
maxlen: StreamMaxlen,
id: ID,
items: &'a [(F, V)]
) -> &mut Self

Add a stream message while capping the stream at a maxlength.

XADD key [MAXLEN [~|=] <count>] <ID or *> [field value] [field value] ...

pub fn xadd_maxlen_map<'a, K: ToRedisArgs, ID: ToRedisArgs, BTM: ToRedisArgs>(
&mut self,
key: K,
maxlen: StreamMaxlen,
id: ID,
map: BTM
) -> &mut Self

BTreeMap variant for adding a stream message while capping the stream at a maxlength.

XADD key [MAXLEN [~|=] <count>] <ID or *> [rust BTreeMap] ...

pub fn xclaim<'a, K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs, MIT: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
key: K,
group: G,
consumer: C,
min_idle_time: MIT,
ids: &'a [ID]
) -> &mut Self

Claim pending, unacked messages, after some period of time, currently checked out by another consumer.

This method only accepts the must-have arguments for claiming messages. If optional arguments are required, see xclaim_options below.

XCLAIM <key> <group> <consumer> <min-idle-time> [<ID-1> <ID-2>]

pub fn xclaim_options<'a, K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs, MIT: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
key: K,
group: G,
consumer: C,
min_idle_time: MIT,
ids: &'a [ID],
options: StreamClaimOptions
) -> &mut Self

This is the optional arguments version for claiming unacked, pending messages currently checked out by another consumer.

use redis::{Connection,Commands,RedisResult};
use redis::streams::{StreamClaimOptions,StreamClaimReply};
let client = redis::Client::open("redis://127.0.0.1/0").unwrap();
let mut con = client.get_connection().unwrap();

// Claim all pending messages for key "k1",
// from group "g1", checked out by consumer "c1"
// for 10ms with RETRYCOUNT 2 and FORCE

let opts = StreamClaimOptions::default()
    .with_force()
    .retry(2);
let results: RedisResult<StreamClaimReply> =
    con.xclaim_options("k1", "g1", "c1", 10, &["0"], opts);

// All optional arguments return a `Result<StreamClaimReply>` with one exception:
// Passing JUSTID returns only the message `id` and omits the HashMap for each message.

let opts = StreamClaimOptions::default()
    .with_justid();
let results: RedisResult<Vec<String>> =
    con.xclaim_options("k1", "g1", "c1", 10, &["0"], opts);

XCLAIM <key> <group> <consumer> <min-idle-time> <ID-1> <ID-2>
    [IDLE <milliseconds>] [TIME <mstime>] [RETRYCOUNT <count>]
    [FORCE] [JUSTID]

pub fn xdel<'a, K: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
key: K,
ids: &'a [ID]
) -> &mut Self

Deletes a list of ids for a given stream key.

XDEL <key> [<ID1> <ID2> ... <IDN>]

pub fn xgroup_create<'a, K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
key: K,
group: G,
id: ID
) -> &mut Self

This command is used for creating a consumer group. It expects the stream key to already exist. Otherwise, use xgroup_create_mkstream if it doesn’t. The id is the starting message id all consumers should read from. Use $ If you want all consumers to read from the last message added to stream.

XGROUP CREATE <key> <groupname> <id or $>

pub fn xgroup_create_mkstream<'a, K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
key: K,
group: G,
id: ID
) -> &mut Self

This is the alternate version for creating a consumer group which makes the stream if it doesn’t exist.

XGROUP CREATE <key> <groupname> <id or $> [MKSTREAM]

pub fn xgroup_setid<'a, K: ToRedisArgs, G: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
key: K,
group: G,
id: ID
) -> &mut Self

Alter which id you want consumers to begin reading from an existing consumer group.

XGROUP SETID <key> <groupname> <id or $>

pub fn xgroup_destroy<'a, K: ToRedisArgs, G: ToRedisArgs>(
&mut self,
key: K,
group: G
) -> &mut Self

Destroy an existing consumer group for a given stream key

XGROUP SETID <key> <groupname> <id or $>

pub fn xgroup_delconsumer<'a, K: ToRedisArgs, G: ToRedisArgs, C: ToRedisArgs>(
&mut self,
key: K,
group: G,
consumer: C
) -> &mut Self

This deletes a consumer from an existing consumer group for given stream `key.

XGROUP DELCONSUMER <key> <groupname> <consumername>

pub fn xinfo_consumers<'a, K: ToRedisArgs, G: ToRedisArgs>(
&mut self,
key: K,
group: G
) -> &mut Self

This returns all info details about which consumers have read messages for given consumer group. Take note of the StreamInfoConsumersReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO CONSUMERS <key> <group>

pub fn xinfo_groups<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Returns all consumer groups created for a given stream key. Take note of the StreamInfoGroupsReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO GROUPS <key>

pub fn xinfo_stream<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Returns info about high-level stream details (first & last message id, length, number of groups, etc.) Take note of the StreamInfoStreamReply return type.

It’s possible this return value might not contain new fields added by Redis in future versions.

XINFO STREAM <key>

pub fn xlen<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

Returns the number of messages for a given stream key.

XLEN <key>

pub fn xpending<'a, K: ToRedisArgs, G: ToRedisArgs>(
&mut self,
key: K,
group: G
) -> &mut Self

This is a basic version of making XPENDING command calls which only passes a stream key and consumer group and it returns details about which consumers have pending messages that haven’t been acked.

You can use this method along with xclaim or xclaim_options for determining which messages need to be retried.

Take note of the StreamPendingReply return type.

XPENDING <key> <group> [<start> <stop> <count> [<consumer>]]

pub fn xpending_count<'a, K: ToRedisArgs, G: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs>(
&mut self,
key: K,
group: G,
start: S,
end: E,
count: C
) -> &mut Self

This XPENDING version returns a list of all messages over the range. You can use this for paginating pending messages (but without the message HashMap).

Start and end follow the same rules xrange args. Set start to - and end to + for the entire stream.

Take note of the StreamPendingCountReply return type.

XPENDING <key> <group> <start> <stop> <count>

pub fn xpending_consumer_count<'a, K: ToRedisArgs, G: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs, CN: ToRedisArgs>(
&mut self,
key: K,
group: G,
start: S,
end: E,
count: C,
consumer: CN
) -> &mut Self

An alternate version of xpending_count which filters by consumer name.

Start and end follow the same rules xrange args. Set start to - and end to + for the entire stream.

Take note of the StreamPendingCountReply return type.

XPENDING <key> <group> <start> <stop> <count> <consumer>

pub fn xrange<'a, K: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs>(
&mut self,
key: K,
start: S,
end: E
) -> &mut Self

Returns a range of messages in a given stream key.

Set start to - to begin at the first message. Set end to + to end the most recent message. You can pass message id to both start and end.

Take note of the StreamRangeReply return type.

XRANGE key start end

pub fn xrange_all<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

A helper method for automatically returning all messages in a stream by key. Use with caution!

XRANGE key - +

pub fn xrange_count<'a, K: ToRedisArgs, S: ToRedisArgs, E: ToRedisArgs, C: ToRedisArgs>(
&mut self,
key: K,
start: S,
end: E,
count: C
) -> &mut Self

A method for paginating a stream by key.

XRANGE key start end [COUNT <n>]

pub fn xread<'a, K: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
keys: &'a [K],
ids: &'a [ID]
) -> &mut Self

Read a list of ids for each stream key. This is the basic form of reading streams. For more advanced control, like blocking, limiting, or reading by consumer group, see xread_options.

XREAD STREAMS key_1 key_2 ... key_N ID_1 ID_2 ... ID_N

pub fn xread_options<'a, K: ToRedisArgs, ID: ToRedisArgs>(
&mut self,
keys: &'a [K],
ids: &'a [ID],
options: StreamReadOptions
) -> &mut Self

This method handles setting optional arguments for XREAD or XREADGROUP Redis commands.

use redis::{Connection,RedisResult,Commands};
use redis::streams::{StreamReadOptions,StreamReadReply};
let client = redis::Client::open("redis://127.0.0.1/0").unwrap();
let mut con = client.get_connection().unwrap();

// Read 10 messages from the start of the stream,
// without registering as a consumer group.

let opts = StreamReadOptions::default()
    .count(10);
let results: RedisResult<StreamReadReply> =
    con.xread_options(&["k1"], &["0"], opts);

// Read all undelivered messages for a given
// consumer group. Be advised: the consumer group must already
// exist before making this call. Also note: we're passing
// '>' as the id here, which means all undelivered messages.

let opts = StreamReadOptions::default()
    .group("group-1", "consumer-1");
let results: RedisResult<StreamReadReply> =
    con.xread_options(&["k1"], &[">"], opts);

XREAD [BLOCK <milliseconds>] [COUNT <count>]
    STREAMS key_1 key_2 ... key_N
    ID_1 ID_2 ... ID_N

XREADGROUP [BLOCK <milliseconds>] [COUNT <count>] [NOACK] [GROUP group-name consumer-name]
    STREAMS key_1 key_2 ... key_N
    ID_1 ID_2 ... ID_N

pub fn xrevrange<'a, K: ToRedisArgs, E: ToRedisArgs, S: ToRedisArgs>(
&mut self,
key: K,
end: E,
start: S
) -> &mut Self

This is the reverse version of xrange. The same rules apply for start and end here.

XREVRANGE key end start

pub fn xrevrange_all<'a, K: ToRedisArgs>(&mut self, key: K) -> &mut Self

This is the reverse version of xrange_all. The same rules apply for start and end here.

XREVRANGE key + -

pub fn xrevrange_count<'a, K: ToRedisArgs, E: ToRedisArgs, S: ToRedisArgs, C: ToRedisArgs>(
&mut self,
key: K,
end: E,
start: S,
count: C
) -> &mut Self

This is the reverse version of xrange_count. The same rules apply for start and end here.

XREVRANGE key end start [COUNT <n>]

pub fn xtrim<'a, K: ToRedisArgs>(
&mut self,
key: K,
maxlen: StreamMaxlen
) -> &mut Self

Trim a stream key to a MAXLEN count.

XTRIM <key> MAXLEN [~|=] <count>  (Same as XADD MAXLEN option)