Struct redis::Pipeline

pub struct Pipeline { /* fields omitted */ }

Represents a redis command pipeline.

Methods

impl Pipeline

A pipeline allows you to send multiple commands in one go to the redis server. API wise it's very similar to just using a command but it allows multiple commands to be chained and some features such as iteration are not available.

Basic example:

let ((k1, k2),) : ((i32, i32),) = redis::pipe()
    .cmd("SET").arg("key_1").arg(42).ignore()
    .cmd("SET").arg("key_2").arg(43).ignore()
    .cmd("MGET").arg(&["key_1", "key_2"]).query(&con).unwrap();

As you can see with cmd you can start a new command. By default each command produces a value but for some you can ignore them by calling ignore on the command. That way it will be skipped in the return value which is useful for SET commands and others, which do not have a useful return value.

pub fn new() -> Pipeline

Creates an empty pipeline. For consistency with the cmd api a pipe function is provided as alias.

pub fn cmd(&mut self, name: &str) -> &mut Pipeline

Starts a new command. Functions such as arg then become available to add more arguments to that command.

pub fn add_command(&mut self, cmd: &Cmd) -> &mut Pipeline

Adds a command to the pipeline.

pub fn arg<T: ToRedisArgs>(&mut self, arg: T) -> &mut Pipeline

Adds an argument to the last started command. This works similar to the arg method of the Cmd object.

Note that this function fails the task if executed on an empty pipeline.

pub fn ignore(&mut self) -> &mut Pipeline

Instructs the pipeline to ignore the return value of this command. It will still be ensured that it is not an error, but any successful result is just thrown away. This makes result processing through tuples much easier because you do not need to handle all the items you do not care about.

Note that this function fails the task if executed on an empty pipeline.

pub fn atomic(&mut self) -> &mut Pipeline

This enables atomic mode. In atomic mode the whole pipeline is enclosed in MULTI/EXEC. From the user's point of view nothing changes however. This is easier than using MULTI/EXEC yourself as the format does not change.

let (k1, k2) : (i32, i32) = redis::pipe()
    .atomic()
    .cmd("GET").arg("key_1")
    .cmd("GET").arg("key_2").query(&con).unwrap();

pub fn query<T: FromRedisValue>(
&self,
con: &dyn ConnectionLike
) -> RedisResult<T>

Executes the pipeline and fetches the return values. Since most pipelines return different types it's recommended to use tuple matching to process the results:

let (k1, k2) : (i32, i32) = redis::pipe()
    .cmd("SET").arg("key_1").arg(42).ignore()
    .cmd("SET").arg("key_2").arg(43).ignore()
    .cmd("GET").arg("key_1")
    .cmd("GET").arg("key_2").query(&con).unwrap();

pub fn query_async<C, T: FromRedisValue>(self, con: C) -> RedisFuture<(C, T)> where
C: ConnectionLike + Send + 'static,
T: Send + 'static,

pub fn execute(&self, con: &dyn ConnectionLike)

This is a shortcut to query() that does not return a value and will fail the task if the query of the pipeline fails.

This is equivalent to a call of query like this:

let _ : () = redis::pipe().cmd("PING").query(&con).unwrap();