redis

Struct PubSub

pub struct PubSub<'a> { /* private fields */ }

Represents a RESP2 pubsub connection.

If you’re using a DB that supports RESP3, consider using a regular connection and setting a push sender it using Connection::set_push_sender.

Implementations

impl<'a> PubSub<'a>

The pubsub object provides convenient access to the redis pubsub system. Once created you can subscribe and unsubscribe from channels and listen in on messages.

Example:

let client = redis::Client::open("redis://127.0.0.1/")?;
let mut con = client.get_connection()?;
let mut pubsub = con.as_pubsub();
pubsub.subscribe("channel_1")?;
pubsub.subscribe("channel_2")?;

loop {
    let msg = pubsub.get_message()?;
    let payload : String = msg.get_payload()?;
    println!("channel '{}': {}", msg.get_channel_name(), payload);
}

pub fn subscribe<T: ToRedisArgs>(&mut self, channel: T) -> RedisResult<()>

Subscribes to a new channel(s).

pub fn psubscribe<T: ToRedisArgs>(&mut self, pchannel: T) -> RedisResult<()>

Subscribes to new channel(s) with pattern(s).

pub fn unsubscribe<T: ToRedisArgs>(&mut self, channel: T) -> RedisResult<()>

Unsubscribes from a channel(s).

pub fn punsubscribe<T: ToRedisArgs>(&mut self, pchannel: T) -> RedisResult<()>

Unsubscribes from channel pattern(s).

pub fn ping_message<T: FromRedisValue>( &mut self, message: impl ToRedisArgs, ) -> RedisResult<T>

Sends a ping with a message to the server

pub fn ping<T: FromRedisValue>(&mut self) -> RedisResult<T>

Sends a ping to the server

pub fn get_message(&mut self) -> RedisResult<Msg>

Fetches the next message from the pubsub connection. Blocks until a message becomes available. This currently does not provide a wait not to block :(

The message itself is still generic and can be converted into an appropriate type through the helper methods on it.

pub fn set_read_timeout(&self, dur: Option<Duration>) -> RedisResult<()>

Sets the read timeout for the connection.

If the provided value is None, then get_message call will block indefinitely. It is an error to pass the zero Duration to this method.