Struct redis::RedisErrorExperimental

pub struct RedisError {
    pub kind: ErrorKind,
    pub desc: &'static str,
    pub detail: Option<String>,
}

Represents a redis error. For the most part you should be using the Error trait to interact with this rather than the actual struct.

Fields

kind

desc

detail

Methods

impl RedisError

Indicates a general failure in the library.

fn category(&self) -> &str

Returns the name of the error category for display purposes.

fn is_io_error(&self) -> bool

Indicates that this failure is an IO failure.

fn is_connection_refusal(&self) -> bool

Returns true if this error indicates that the connection was refused.