Trait redis::Commands

pub trait Commands: ConnectionLike + Sized {
    fn get<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn keys<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn set<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn set_multiple<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        items: &'a [(K, V)]
    ) -> RedisResult<RV> { ... }

    fn set_ex<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V, 
        seconds: usize
    ) -> RedisResult<RV> { ... }

    fn pset_ex<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V, 
        milliseconds: usize
    ) -> RedisResult<RV> { ... }

    fn set_nx<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn mset_nx<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        items: &'a [(K, V)]
    ) -> RedisResult<RV> { ... }

    fn getset<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn getrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        from: isize, 
        to: isize
    ) -> RedisResult<RV> { ... }

    fn setrange<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        offset: isize, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn del<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn exists<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn expire<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        seconds: usize
    ) -> RedisResult<RV> { ... }

    fn expire_at<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        ts: usize
    ) -> RedisResult<RV> { ... }

    fn pexpire<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        ms: usize
    ) -> RedisResult<RV> { ... }

    fn pexpire_at<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        ts: usize
    ) -> RedisResult<RV> { ... }

    fn persist<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn ttl<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn pttl<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn rename<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        new_key: K
    ) -> RedisResult<RV> { ... }

    fn rename_nx<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        new_key: K
    ) -> RedisResult<RV> { ... }

    fn append<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn incr<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        delta: V
    ) -> RedisResult<RV> { ... }

    fn setbit<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        offset: usize, 
        value: bool
    ) -> RedisResult<RV> { ... }

    fn getbit<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        offset: usize
    ) -> RedisResult<RV> { ... }

    fn bitcount<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn bitcount_range<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: usize, 
        end: usize
    ) -> RedisResult<RV> { ... }

    fn bit_and<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        srckeys: K
    ) -> RedisResult<RV> { ... }

    fn bit_or<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        srckeys: K
    ) -> RedisResult<RV> { ... }

    fn bit_xor<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        srckeys: K
    ) -> RedisResult<RV> { ... }

    fn bit_not<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        srckey: K
    ) -> RedisResult<RV> { ... }

    fn strlen<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn hget<'a, K: ToRedisArgs, F: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        field: F
    ) -> RedisResult<RV> { ... }

    fn hdel<'a, K: ToRedisArgs, F: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        field: F
    ) -> RedisResult<RV> { ... }

    fn hset<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        field: F, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn hset_nx<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        field: F, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn hset_multiple<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        items: &'a [(F, V)]
    ) -> RedisResult<RV> { ... }

    fn hincr<'a, K: ToRedisArgs, F: ToRedisArgs, D: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        field: F, 
        delta: D
    ) -> RedisResult<RV> { ... }

    fn hexists<'a, K: ToRedisArgs, F: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        field: F
    ) -> RedisResult<RV> { ... }

    fn hkeys<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn hvals<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn hgetall<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn hlen<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn blpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        timeout: usize
    ) -> RedisResult<RV> { ... }

    fn brpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        timeout: usize
    ) -> RedisResult<RV> { ... }

    fn brpoplpush<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        srckey: K, 
        dstkey: K, 
        timeout: usize
    ) -> RedisResult<RV> { ... }

    fn lindex<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        index: isize
    ) -> RedisResult<RV> { ... }

    fn linsert_before<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        pivot: P, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn linsert_after<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        pivot: P, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn llen<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn lpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn lpush<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn lpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn lrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn lrem<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        count: isize, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn ltrim<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn lset<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        index: isize, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn rpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn rpoplpush<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        dstkey: K
    ) -> RedisResult<RV> { ... }

    fn rpush<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn rpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        value: V
    ) -> RedisResult<RV> { ... }

    fn sadd<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn scard<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn sdiff<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        keys: K
    ) -> RedisResult<RV> { ... }

    fn sdiffstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: K
    ) -> RedisResult<RV> { ... }

    fn sinter<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        keys: K
    ) -> RedisResult<RV> { ... }

    fn sinterstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: K
    ) -> RedisResult<RV> { ... }

    fn sismember<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn smembers<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn smove<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        srckey: K, 
        dstkey: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn spop<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn srandmember<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn srandmember_multiple<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        count: usize
    ) -> RedisResult<RV> { ... }

    fn srem<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn sunion<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        keys: K
    ) -> RedisResult<RV> { ... }

    fn sunionstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: K
    ) -> RedisResult<RV> { ... }

    fn zadd<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M, 
        score: S
    ) -> RedisResult<RV> { ... }

    fn zadd_multiple<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        items: &'a [(S, M)]
    ) -> RedisResult<RV> { ... }

    fn zcard<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn zcount<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM
    ) -> RedisResult<RV> { ... }

    fn zincr<'a, K: ToRedisArgs, M: ToRedisArgs, D: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M, 
        delta: D
    ) -> RedisResult<RV> { ... }

    fn zinterstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: &'a [K]
    ) -> RedisResult<RV> { ... }

    fn zinterstore_min<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: &'a [K]
    ) -> RedisResult<RV> { ... }

    fn zinterstore_max<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: &'a [K]
    ) -> RedisResult<RV> { ... }

    fn zlexcount<'a, K: ToRedisArgs, L: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: L, 
        max: L
    ) -> RedisResult<RV> { ... }

    fn zrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn zrange_withscores<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn zrangebylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM
    ) -> RedisResult<RV> { ... }

    fn zrangebylex_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM, 
        offset: isize, 
        count: isize
    ) -> RedisResult<RV> { ... }

    fn zrevrangebylex<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        max: MM, 
        min: M
    ) -> RedisResult<RV> { ... }

    fn zrevrangebylex_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        max: MM, 
        min: M, 
        offset: isize, 
        count: isize
    ) -> RedisResult<RV> { ... }

    fn zrangebyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM
    ) -> RedisResult<RV> { ... }

    fn zrangebyscore_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM
    ) -> RedisResult<RV> { ... }

    fn zrangebyscore_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM, 
        offset: isize, 
        count: isize
    ) -> RedisResult<RV> { ... }

    fn zrangebyscore_limit_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM, 
        offset: isize, 
        count: isize
    ) -> RedisResult<RV> { ... }

    fn zrank<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn zrem<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        members: M
    ) -> RedisResult<RV> { ... }

    fn zrembylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM
    ) -> RedisResult<RV> { ... }

    fn zremrangebyrank<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn zrembyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        min: M, 
        max: MM
    ) -> RedisResult<RV> { ... }

    fn zrevrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn zrevrange_withscores<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        start: isize, 
        stop: isize
    ) -> RedisResult<RV> { ... }

    fn zrevrangebyscore<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        max: MM, 
        min: M
    ) -> RedisResult<RV> { ... }

    fn zrevrangebyscore_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        max: MM, 
        min: M
    ) -> RedisResult<RV> { ... }

    fn zrevrangebyscore_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        max: MM, 
        min: M, 
        offset: isize, 
        count: isize
    ) -> RedisResult<RV> { ... }

    fn zrevrangebyscore_limit_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        max: MM, 
        min: M, 
        offset: isize, 
        count: isize
    ) -> RedisResult<RV> { ... }

    fn zrevrank<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn zscore<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M
    ) -> RedisResult<RV> { ... }

    fn zunionstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: &'a [K]
    ) -> RedisResult<RV> { ... }

    fn zunionstore_min<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: &'a [K]
    ) -> RedisResult<RV> { ... }

    fn zunionstore_max<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        keys: &'a [K]
    ) -> RedisResult<RV> { ... }

    fn pfadd<'a, K: ToRedisArgs, E: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        element: E
    ) -> RedisResult<RV> { ... }

    fn pfcount<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<RV> { ... }

    fn pfmerge<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        dstkey: K, 
        srckeys: K
    ) -> RedisResult<RV> { ... }

    fn publish<'a, K: ToRedisArgs, E: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        channel: K, 
        message: E
    ) -> RedisResult<RV> { ... }

    fn geo_add<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        members: M
    ) -> RedisResult<RV> { ... }

    fn geo_dist<'a, K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member1: M1, 
        member2: M2, 
        unit: Unit
    ) -> RedisResult<RV> { ... }

    fn geo_hash<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        members: M
    ) -> RedisResult<RV> { ... }

    fn geo_pos<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        members: M
    ) -> RedisResult<RV> { ... }

    fn geo_radius<'a, K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        longitude: f64, 
        latitude: f64, 
        radius: f64, 
        unit: Unit, 
        options: RadiusOptions
    ) -> RedisResult<RV> { ... }

    fn geo_radius_by_member<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        member: M, 
        radius: f64, 
        unit: Unit, 
        options: RadiusOptions
    ) -> RedisResult<RV> { ... }

    fn scan<RV: FromRedisValue>(&mut self) -> RedisResult<Iter<RV>> { ... }

    fn scan_match<P: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        pattern: P
    ) -> RedisResult<Iter<RV>> { ... }

    fn hscan<K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<Iter<RV>> { ... }

    fn hscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        pattern: P
    ) -> RedisResult<Iter<RV>> { ... }

    fn sscan<K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<Iter<RV>> { ... }

    fn sscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        pattern: P
    ) -> RedisResult<Iter<RV>> { ... }

    fn zscan<K: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K
    ) -> RedisResult<Iter<RV>> { ... }

    fn zscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>(
        &mut self, 
        key: K, 
        pattern: P
    ) -> RedisResult<Iter<RV>> { ... }
}

Implements common redis commands for connection like objects. This allows you to send commands straight to a connection or client. It is also implemented for redis results of clients which makes for very convenient access in some basic cases.

This allows you to use nicer syntax for some common operations. For instance this code:

let client = redis::Client::open("redis://127.0.0.1/")?;
let mut con = client.get_connection()?;
redis::cmd("SET").arg("my_key").arg(42).execute(&mut con);
assert_eq!(redis::cmd("GET").arg("my_key").query(&mut con), Ok(42));

Will become this:

use redis::Commands;
let client = redis::Client::open("redis://127.0.0.1/")?;
let mut con = client.get_connection()?;
let () = con.set("my_key", 42)?;
assert_eq!(con.get("my_key"), Ok(42));

Provided methods

fn get<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get the value of a key. If key is a vec this becomes an MGET.

fn keys<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Gets all keys matching pattern

fn set<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Set the string value of a key.

fn set_multiple<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
items: &'a [(K, V)]
) -> RedisResult<RV>

Sets multiple keys to their values.

fn set_ex<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V,
seconds: usize
) -> RedisResult<RV>

Set the value and expiration of a key.

fn pset_ex<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V,
milliseconds: usize
) -> RedisResult<RV>

Set the value and expiration in milliseconds of a key.

fn set_nx<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Set the value of a key, only if the key does not exist

fn mset_nx<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
items: &'a [(K, V)]
) -> RedisResult<RV>

Sets multiple keys to their values failing if at least one already exists.

fn getset<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Set the string value of a key and return its old value.

fn getrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
from: isize,
to: isize
) -> RedisResult<RV>

Get a range of bytes/substring from the value of a key. Negative values provide an offset from the end of the value.

fn setrange<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
offset: isize,
value: V
) -> RedisResult<RV>

Overwrite the part of the value stored in key at the specified offset.

fn del<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Delete one or more keys.

fn exists<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Determine if a key exists.

fn expire<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
seconds: usize
) -> RedisResult<RV>

Set a key's time to live in seconds.

fn expire_at<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
ts: usize
) -> RedisResult<RV>

Set the expiration for a key as a UNIX timestamp.

fn pexpire<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
ms: usize
) -> RedisResult<RV>

Set a key's time to live in milliseconds.

fn pexpire_at<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
ts: usize
) -> RedisResult<RV>

Set the expiration for a key as a UNIX timestamp in milliseconds.

fn persist<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Remove the expiration from a key.

fn ttl<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get the expiration time of a key.

fn pttl<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get the expiration time of a key in milliseconds.

fn rename<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
new_key: K
) -> RedisResult<RV>

Rename a key.

fn rename_nx<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
new_key: K
) -> RedisResult<RV>

Rename a key, only if the new key does not exist.

fn append<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Append a value to a key.

fn incr<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
delta: V
) -> RedisResult<RV>

Increment the numeric value of a key by the given amount. This issues a INCRBY or INCRBYFLOAT depending on the type.

fn setbit<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
offset: usize,
value: bool
) -> RedisResult<RV>

Sets or clears the bit at offset in the string value stored at key.

fn getbit<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
offset: usize
) -> RedisResult<RV>

Returns the bit value at offset in the string value stored at key.

fn bitcount<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Count set bits in a string.

fn bitcount_range<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: usize,
end: usize
) -> RedisResult<RV>

Count set bits in a string in a range.

fn bit_and<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
srckeys: K
) -> RedisResult<RV>

Perform a bitwise AND between multiple keys (containing string values) and store the result in the destination key.

fn bit_or<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
srckeys: K
) -> RedisResult<RV>

Perform a bitwise OR between multiple keys (containing string values) and store the result in the destination key.

fn bit_xor<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
srckeys: K
) -> RedisResult<RV>

Perform a bitwise XOR between multiple keys (containing string values) and store the result in the destination key.

fn bit_not<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
srckey: K
) -> RedisResult<RV>

Perform a bitwise NOT of the key (containing string values) and store the result in the destination key.

fn strlen<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get the length of the value stored in a key.

fn hget<'a, K: ToRedisArgs, F: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
field: F
) -> RedisResult<RV>

Gets a single (or multiple) fields from a hash.

fn hdel<'a, K: ToRedisArgs, F: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
field: F
) -> RedisResult<RV>

Deletes a single (or multiple) fields from a hash.

fn hset<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
field: F,
value: V
) -> RedisResult<RV>

Sets a single field in a hash.

fn hset_nx<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
field: F,
value: V
) -> RedisResult<RV>

Sets a single field in a hash if it does not exist.

fn hset_multiple<'a, K: ToRedisArgs, F: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
items: &'a [(F, V)]
) -> RedisResult<RV>

Sets a multiple fields in a hash.

fn hincr<'a, K: ToRedisArgs, F: ToRedisArgs, D: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
field: F,
delta: D
) -> RedisResult<RV>

Increments a value.

fn hexists<'a, K: ToRedisArgs, F: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
field: F
) -> RedisResult<RV>

Checks if a field in a hash exists.

fn hkeys<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Gets all the keys in a hash.

fn hvals<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Gets all the values in a hash.

fn hgetall<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Gets all the fields and values in a hash.

fn hlen<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Gets the length of a hash.

fn blpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
timeout: usize
) -> RedisResult<RV>

Remove and get the first element in a list, or block until one is available.

fn brpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
timeout: usize
) -> RedisResult<RV>

Remove and get the last element in a list, or block until one is available.

fn brpoplpush<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
srckey: K,
dstkey: K,
timeout: usize
) -> RedisResult<RV>

Pop a value from a list, push it to another list and return it; or block until one is available.

fn lindex<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
index: isize
) -> RedisResult<RV>

Get an element from a list by its index.

fn linsert_before<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
pivot: P,
value: V
) -> RedisResult<RV>

Insert an element before another element in a list.

fn linsert_after<'a, K: ToRedisArgs, P: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
pivot: P,
value: V
) -> RedisResult<RV>

Insert an element after another element in a list.

fn llen<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Returns the length of the list stored at key.

fn lpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Removes and returns the first element of the list stored at key.

fn lpush<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Insert all the specified values at the head of the list stored at key.

fn lpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Inserts a value at the head of the list stored at key, only if key already exists and holds a list.

fn lrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Returns the specified elements of the list stored at key.

fn lrem<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
count: isize,
value: V
) -> RedisResult<RV>

Removes the first count occurrences of elements equal to value from the list stored at key.

fn ltrim<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Trim an existing list so that it will contain only the specified range of elements specified.

fn lset<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
index: isize,
value: V
) -> RedisResult<RV>

Sets the list element at index to value

fn rpop<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Removes and returns the last element of the list stored at key.

fn rpoplpush<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
dstkey: K
) -> RedisResult<RV>

Pop a value from a list, push it to another list and return it.

fn rpush<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Insert all the specified values at the tail of the list stored at key.

fn rpush_exists<'a, K: ToRedisArgs, V: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
value: V
) -> RedisResult<RV>

Inserts value at the tail of the list stored at key, only if key already exists and holds a list.

fn sadd<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M
) -> RedisResult<RV>

Add one or more members to a set.

fn scard<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get the number of members in a set.

fn sdiff<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
keys: K
) -> RedisResult<RV>

Subtract multiple sets.

fn sdiffstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: K
) -> RedisResult<RV>

Subtract multiple sets and store the resulting set in a key.

fn sinter<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
keys: K
) -> RedisResult<RV>

Intersect multiple sets.

fn sinterstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: K
) -> RedisResult<RV>

Intersect multiple sets and store the resulting set in a key.

fn sismember<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M
) -> RedisResult<RV>

Determine if a given value is a member of a set.

fn smembers<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get all the members in a set.

fn smove<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
srckey: K,
dstkey: K,
member: M
) -> RedisResult<RV>

Move a member from one set to another.

fn spop<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Remove and return a random member from a set.

fn srandmember<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get one random member from a set.

fn srandmember_multiple<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
count: usize
) -> RedisResult<RV>

Get multiple random members from a set.

fn srem<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M
) -> RedisResult<RV>

Remove one or more members from a set.

fn sunion<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
keys: K
) -> RedisResult<RV>

Add multiple sets.

fn sunionstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: K
) -> RedisResult<RV>

Add multiple sets and store the resulting set in a key.

fn zadd<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M,
score: S
) -> RedisResult<RV>

Add one member to a sorted set, or update its score if it already exists.

fn zadd_multiple<'a, K: ToRedisArgs, S: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
items: &'a [(S, M)]
) -> RedisResult<RV>

Add multiple members to a sorted set, or update its score if it already exists.

fn zcard<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Get the number of members in a sorted set.

fn zcount<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM
) -> RedisResult<RV>

Count the members in a sorted set with scores within the given values.

fn zincr<'a, K: ToRedisArgs, M: ToRedisArgs, D: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M,
delta: D
) -> RedisResult<RV>

Increments the member in a sorted set at key by delta. If the member does not exist, it is added with delta as its score.

fn zinterstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> RedisResult<RV>

Intersect multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

fn zinterstore_min<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> RedisResult<RV>

Intersect multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

fn zinterstore_max<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> RedisResult<RV>

Intersect multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

fn zlexcount<'a, K: ToRedisArgs, L: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: L,
max: L
) -> RedisResult<RV>

Count the number of members in a sorted set between a given lexicographical range.

fn zrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by index

fn zrange_withscores<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by index with scores.

fn zrangebylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM
) -> RedisResult<RV>

Return a range of members in a sorted set, by lexicographical range.

fn zrangebylex_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by lexicographical range with offset and limit.

fn zrevrangebylex<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
max: MM,
min: M
) -> RedisResult<RV>

Return a range of members in a sorted set, by lexicographical range.

fn zrevrangebylex_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by lexicographical range with offset and limit.

fn zrangebyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM
) -> RedisResult<RV>

Return a range of members in a sorted set, by score.

fn zrangebyscore_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM
) -> RedisResult<RV>

Return a range of members in a sorted set, by score with scores.

fn zrangebyscore_limit<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by score with limit.

fn zrangebyscore_limit_withscores<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM,
offset: isize,
count: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by score with limit with scores.

fn zrank<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M
) -> RedisResult<RV>

Determine the index of a member in a sorted set.

fn zrem<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
members: M
) -> RedisResult<RV>

Remove one or more members from a sorted set.

fn zrembylex<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM
) -> RedisResult<RV>

Remove all members in a sorted set between the given lexicographical range.

fn zremrangebyrank<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Remove all members in a sorted set within the given indexes.

fn zrembyscore<'a, K: ToRedisArgs, M: ToRedisArgs, MM: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
min: M,
max: MM
) -> RedisResult<RV>

Remove all members in a sorted set within the given scores.

fn zrevrange<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by index, with scores ordered from high to low.

fn zrevrange_withscores<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
start: isize,
stop: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by index, with scores ordered from high to low.

fn zrevrangebyscore<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
max: MM,
min: M
) -> RedisResult<RV>

Return a range of members in a sorted set, by score.

fn zrevrangebyscore_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
max: MM,
min: M
) -> RedisResult<RV>

Return a range of members in a sorted set, by score with scores.

fn zrevrangebyscore_limit<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by score with limit.

fn zrevrangebyscore_limit_withscores<'a, K: ToRedisArgs, MM: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
max: MM,
min: M,
offset: isize,
count: isize
) -> RedisResult<RV>

Return a range of members in a sorted set, by score with limit with scores.

fn zrevrank<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M
) -> RedisResult<RV>

Determine the index of a member in a sorted set, with scores ordered from high to low.

fn zscore<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M
) -> RedisResult<RV>

Get the score associated with the given member in a sorted set.

fn zunionstore<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> RedisResult<RV>

Unions multiple sorted sets and store the resulting sorted set in a new key using SUM as aggregation function.

fn zunionstore_min<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> RedisResult<RV>

Unions multiple sorted sets and store the resulting sorted set in a new key using MIN as aggregation function.

fn zunionstore_max<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
keys: &'a [K]
) -> RedisResult<RV>

Unions multiple sorted sets and store the resulting sorted set in a new key using MAX as aggregation function.

fn pfadd<'a, K: ToRedisArgs, E: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
element: E
) -> RedisResult<RV>

Adds the specified elements to the specified HyperLogLog.

fn pfcount<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<RV>

Return the approximated cardinality of the set(s) observed by the HyperLogLog at key(s).

fn pfmerge<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
dstkey: K,
srckeys: K
) -> RedisResult<RV>

Merge N different HyperLogLogs into a single one.

fn publish<'a, K: ToRedisArgs, E: ToRedisArgs, RV: FromRedisValue>(
&mut self,
channel: K,
message: E
) -> RedisResult<RV>

Posts a message to the given channel.

fn geo_add<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
members: M
) -> RedisResult<RV>

Adds the specified geospatial items to the specified key.

Every member has to be written as a tuple of (longitude, latitude, member_name). It can be a single tuple, or a vector of tuples.

longitude, latitude can be set using redis::geo::Coord.

Returns the number of elements added to the sorted set, not including elements already existing for which the score was updated.

Example

use redis::{Commands, Connection, RedisResult};
use redis::geo::Coord;

fn add_point(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", (Coord::lon_lat(13.361389, 38.115556), "Palermo"))
}

fn add_point_with_tuples(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", ("13.361389", "38.115556", "Palermo"))
}

fn add_many_points(con: &mut Connection) -> RedisResult<isize> {
    con.geo_add("my_gis", &[
        ("13.361389", "38.115556", "Palermo"),
        ("15.087269", "37.502669", "Catania")
    ])
}

fn geo_dist<'a, K: ToRedisArgs, M1: ToRedisArgs, M2: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member1: M1,
member2: M2,
unit: Unit
) -> RedisResult<RV>

Return the distance between two members in the geospatial index represented by the sorted set.

If one or both the members are missing, the command returns NULL, so it may be convenient to parse its response as either Option<f64> or Option<String>.

Example

use redis::{Commands, RedisResult};
use redis::geo::Unit;

fn get_dists(con: &mut redis::Connection) {
    let x: RedisResult<f64> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Catania",
        Unit::Kilometers
    );
    // x is Ok(166.2742)

    let x: RedisResult<Option<f64>> = con.geo_dist(
        "my_gis",
        "Palermo",
        "Atlantis",
        Unit::Meters
    );
    // x is Ok(None)
}

fn geo_hash<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
members: M
) -> RedisResult<RV>

Return valid Geohash strings representing the position of one or more members of the geospatial index represented by the sorted set at key.

Example

use redis::{Commands, RedisResult};

fn get_hash(con: &mut redis::Connection) {
    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", "Palermo");
    // x is vec!["sqc8b49rny0"]

    let x: RedisResult<Vec<String>> = con.geo_hash("my_gis", &["Palermo", "Catania"]);
    // x is vec!["sqc8b49rny0", "sqdtr74hyu0"]
}

fn geo_pos<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
members: M
) -> RedisResult<RV>

Return the positions of all the specified members of the geospatial index represented by the sorted set at key.

Every position is a pair of (longitude, latitude). redis::geo::Coord can be used to convert these value in a struct.

Example

use redis::{Commands, RedisResult};
use redis::geo::Coord;

fn get_position(con: &mut redis::Connection) {
    let x: RedisResult<Vec<Vec<f64>>> = con.geo_pos("my_gis", &["Palermo", "Catania"]);
    // x is [ [ 13.361389, 38.115556 ], [ 15.087269, 37.502669 ] ];

    let x: Vec<Coord<f64>> = con.geo_pos("my_gis", "Palermo").unwrap();
    // x[0].longitude is 13.361389
    // x[0].latitude is 38.115556
}

fn geo_radius<'a, K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
longitude: f64,
latitude: f64,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> RedisResult<RV>

Return the members of a sorted set populated with geospatial information using geo_add, which are within the borders of the area specified with the center location and the maximum distance from the center (the radius).

Every item in the result can be read with redis::geo::RadiusSearchResult, which support the multiple formats returned by GEORADIUS.

use redis::{Commands, RedisResult};
use redis::geo::{RadiusOptions, RadiusSearchResult, RadiusOrder, Unit};

fn radius(con: &mut redis::Connection) -> Vec<RadiusSearchResult> {
    let opts = RadiusOptions::default().with_dist().order(RadiusOrder::Asc);
    con.geo_radius("my_gis", 15.90, 37.21, 51.39, Unit::Kilometers, opts).unwrap()
}

fn geo_radius_by_member<'a, K: ToRedisArgs, M: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
member: M,
radius: f64,
unit: Unit,
options: RadiusOptions
) -> RedisResult<RV>

Retrieve members selected by distance with the center of member. The member itself is always contained in the results.

fn scan<RV: FromRedisValue>(&mut self) -> RedisResult<Iter<RV>>

Incrementally iterate the keys space.

fn scan_match<P: ToRedisArgs, RV: FromRedisValue>(
&mut self,
pattern: P
) -> RedisResult<Iter<RV>>

Incrementally iterate the keys space for keys matching a pattern.

fn hscan<K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<Iter<RV>>

Incrementally iterate hash fields and associated values.

fn hscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
pattern: P
) -> RedisResult<Iter<RV>>

Incrementally iterate hash fields and associated values for field names matching a pattern.

fn sscan<K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<Iter<RV>>

Incrementally iterate set elements.

fn sscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
pattern: P
) -> RedisResult<Iter<RV>>

Incrementally iterate set elements for elements matching a pattern.

fn zscan<K: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K
) -> RedisResult<Iter<RV>>

Incrementally iterate sorted set elements.

fn zscan_match<K: ToRedisArgs, P: ToRedisArgs, RV: FromRedisValue>(
&mut self,
key: K,
pattern: P
) -> RedisResult<Iter<RV>>

Incrementally iterate sorted set elements for elements matching a pattern.

Loading content...