Trait redis::IntoConnectionInfoExperimental

pub trait IntoConnectionInfo {
    fn into_connection_info(self) -> RedisResult<ConnectionInfo>;
}

Converts an object into a connection info struct. This allows the constructor of the client to accept connection information in a range of different formats.

Required Methods

fn into_connection_info(self) -> RedisResult<ConnectionInfo>