redis

Trait ToRedisArgs

pub trait ToRedisArgs: Sized {
    // Required method
    fn write_redis_args<W>(&self, out: &mut W)
       where W: ?Sized + RedisWrite;

    // Provided methods
    fn to_redis_args(&self) -> Vec<Vec<u8>> { ... }
    fn describe_numeric_behavior(&self) -> NumericBehavior { ... }
    fn num_of_args(&self) -> usize { ... }
}

Used to convert a value into one or multiple redis argument strings. Most values will produce exactly one item but in some cases it might make sense to produce more than one.

Required Methods

fn write_redis_args<W>(&self, out: &mut W)
where W: ?Sized + RedisWrite,

This writes the value into a vector of bytes. Each item is a single argument. Most items generate a single item.

The exception to this rule currently are vectors of items.

Provided Methods

fn to_redis_args(&self) -> Vec<Vec<u8>>

This converts the value into a vector of bytes. Each item is a single argument. Most items generate a vector of a single item.

The exception to this rule currently are vectors of items.

fn describe_numeric_behavior(&self) -> NumericBehavior

Returns an information about the contained value with regards to it’s numeric behavior in a redis context. This is used in some high level concepts to switch between different implementations of redis functions (for instance INCR vs INCRBYFLOAT).

fn num_of_args(&self) -> usize

Returns the number of arguments this value will generate.

This is used in some high level functions to intelligently switch between GET and MGET variants. Also, for some commands like HEXPIREDAT which require a specific number of arguments, this method can be used to know the number of arguments.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.