Trait redis::ToRedisArgs

pub trait ToRedisArgs: Sized {
    pub fn write_redis_args<W: ?Sized>(&self, out: &mut W)
    where
        W: RedisWrite;

    pub fn to_redis_args(&self) -> Vec<Vec<u8>> { ... }

    pub fn describe_numeric_behavior(&self) -> NumericBehavior { ... }

    pub fn is_single_arg(&self) -> bool { ... }
}

Used to convert a value into one or multiple redis argument strings. Most values will produce exactly one item but in some cases it might make sense to produce more than one.

Required methods

pub fn write_redis_args<W: ?Sized>(&self, out: &mut W) where
W: RedisWrite,

This writes the value into a vector of bytes. Each item is a single argument. Most items generate a single item.

The exception to this rule currently are vectors of items.

Loading content...
Provided methods

pub fn to_redis_args(&self) -> Vec<Vec<u8>>

This converts the value into a vector of bytes. Each item is a single argument. Most items generate a vector of a single item.

The exception to this rule currently are vectors of items.

pub fn describe_numeric_behavior(&self) -> NumericBehavior

Returns an information about the contained value with regards to it's numeric behavior in a redis context. This is used in some high level concepts to switch between different implementations of redis functions (for instance INCR vs INCRBYFLOAT).

pub fn is_single_arg(&self) -> bool

Returns an indiciation if the value contained is exactly one argument. It returns false if it's zero or more than one. This is used in some high level functions to intelligently switch between GET and MGET variants.

Loading content...